Add Coverage Values to Quantile-Based Forecasts

Adds interval coverage of central prediction intervals, quantile coverage for predictive quantiles, as well as the deviation between desired and actual coverage to a data.table. Forecasts should be in a quantile format (following the input requirements of score()).

Interval coverage

Interval coverage for a given interval range is defined as the proportion of observations that fall within the corresponding central prediction intervals. Central prediction intervals are symmetric around the median and and formed by two quantiles that denote the lower and upper bound. For example, the 50% central prediction interval is the interval between the 0.25 and 0.75 quantiles of the predictive distribution.

The function add_coverage() computes the coverage per central prediction interval, so the interval coverage will always be either TRUE (observed value falls within the interval) or FALSE (observed value falls outside the interval). You can summarise the interval coverage values to get the proportion of observations that fall within the central prediction intervals.

Quantile coverage

Quantile coverage for a given quantile is defined as the proportion of observed values that are smaller than the corresponding predictive quantile. For example, the 0.5 quantile coverage is the proportion of observed values that are smaller than the 0.5 quantile of the predictive distribution. Just as above, for a single observation and the quantile of a single predictive distribution, the value will either be TRUE or FALSE.

Coverage deviation

The coverage deviation is the difference between the desired coverage (can be either interval or quantile coverage) and the actual coverage. For example, if the desired coverage is 90% and the actual coverage is 80%, the coverage deviation is -0.1.

Usage

add_coverage(data)

Arguments

data

A data.frame or data.table with predicted and observed values.

Value

a data.table with the input and columns "interval_coverage", "interval_coverage_deviation", "quantile_coverage", "quantile_coverage_deviation" added.

Examples

library(magrittr) # pipe operator
example_quantile %>%
  add_coverage()
#> Some rows containing NA values may be removed. This is fine if not unexpected.
#> Forecast type:
#> [1] "quantile"
#> 
#> Score columns:
#> [1] "interval_coverage"           "interval_coverage_deviation"
#> [3] "quantile_coverage"           "quantile_coverage_deviation"
#> 
#> Forecast unit:
#> [1] "location"        "target_end_date" "target_type"     "location_name"  
#> [5] "forecast_date"   "model"           "horizon"        
#> 
#> Key: <location, target_end_date, target_type, location_name, forecast_date, model, horizon, interval_range>
#>        location target_end_date target_type observed location_name
#>          <char>          <Date>      <char>    <num>        <char>
#>     1:       DE      2021-01-02       Cases   127300       Germany
#>     2:       DE      2021-01-02      Deaths     4534       Germany
#>     3:       DE      2021-01-09       Cases   154922       Germany
#>     4:       DE      2021-01-09      Deaths     6117       Germany
#>     5:       DE      2021-01-16       Cases   110183       Germany
#>    ---                                                            
#> 20541:       IT      2021-07-24      Deaths       78         Italy
#> 20542:       IT      2021-07-24      Deaths       78         Italy
#> 20543:       IT      2021-07-24      Deaths       78         Italy
#> 20544:       IT      2021-07-24      Deaths       78         Italy
#> 20545:       IT      2021-07-24      Deaths       78         Italy
#>        forecast_date quantile_level predicted                model horizon
#>               <Date>          <num>     <int>               <char>   <num>
#>     1:          <NA>             NA        NA                 <NA>      NA
#>     2:          <NA>             NA        NA                 <NA>      NA
#>     3:          <NA>             NA        NA                 <NA>      NA
#>     4:          <NA>             NA        NA                 <NA>      NA
#>     5:          <NA>             NA        NA                 <NA>      NA
#>    ---                                                                    
#> 20541:    2021-07-12          0.950       499 epiforecasts-EpiNow2       2
#> 20542:    2021-07-12          0.025        55 epiforecasts-EpiNow2       2
#> 20543:    2021-07-12          0.975       611 epiforecasts-EpiNow2       2
#> 20544:    2021-07-12          0.010        48 epiforecasts-EpiNow2       2
#> 20545:    2021-07-12          0.990       719 epiforecasts-EpiNow2       2
#>        interval_range interval_coverage interval_coverage_deviation
#>                 <num>            <lgcl>                       <num>
#>     1:             NA                NA                          NA
#>     2:             NA                NA                          NA
#>     3:             NA                NA                          NA
#>     4:             NA                NA                          NA
#>     5:             NA                NA                          NA
#>    ---                                                             
#> 20541:             90              TRUE                        0.10
#> 20542:             95              TRUE                        0.05
#> 20543:             95              TRUE                        0.05
#> 20544:             98              TRUE                        0.02
#> 20545:             98              TRUE                        0.02
#>        quantile_coverage quantile_coverage_deviation
#>                   <lgcl>                       <num>
#>     1:                NA                          NA
#>     2:                NA                          NA
#>     3:                NA                          NA
#>     4:                NA                          NA
#>     5:                NA                          NA
#>    ---                                              
#> 20541:              TRUE                       0.050
#> 20542:             FALSE                      -0.025
#> 20543:              TRUE                       0.025
#> 20544:             FALSE                      -0.010
#> 20545:              TRUE                       0.010