Absolute Error of the Median (Quantile-based Version)

Absolute error of the median calculated as

$$ abs(true_value - median_prediction) $$

ae_median_quantile(true_values, predictions, quantiles = NULL, verbose = TRUE)

Arguments

true_values
A vector with the true observed values of size n

predictions
numeric vector with predictions, corresponding to the quantiles in a second vector, `quantiles`.

quantiles
numeric vector that denotes the quantile for the values in `predictions`. Only those predictions where `quantiles == 0.5` will be kept. If `quantiles` is `NULL`, then all `predictions` and `true_values` will be used (this is then the same as `absolute_error()`)

verbose
logical, return a warning is something unexpected happens

Value

vector with the scoring values

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
ae_median_quantile(true_values, predicted_values, quantiles = 0.5)

#>  [1] 0.65015831 1.72819024 0.06523980 0.20779352 1.66371223 2.64873323
#>  [7] 1.41941848 0.81467725 0.82819755 1.40281054 1.08352273 1.06958300
#> [13] 2.39405728 1.03223771 3.07456881 1.21444768 1.16529004 0.42672765
#> [19] 0.05494542 1.06878221 2.13322599 2.69780322 0.49025398 0.99802378
#> [25] 1.11685712 1.01631557 3.03719277 1.52858776 0.18448801 1.45478936