Absolute Error of the Median (Sample-based Version)

Absolute error of the median calculated as

$$ abs(true_value - median_prediction) $$

ae_median_sample(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n

Value

vector with the scoring values

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
ae_median_sample(true_values, predicted_values)
#>  [1] 0.99372439 0.02581233 1.86127272 0.27273279 1.24653434 1.07227943
#>  [7] 1.23409677 0.08246609 0.59709919 0.91452256 2.90217271 2.63026176
#> [13] 0.18071823 0.22790269 1.39388574 0.08743049 2.38810732 0.69211049
#> [19] 0.48852450 4.71744809 1.77087010 1.08554160 0.53414263 0.60820361
#> [25] 1.18708928 1.57092958 1.95419939 1.67394690 1.24262683 0.20118495