Create a forecast Object

Convert a data.frame or similar of forecasts into an object of class forecast_* and validate it.

as_forecast()

allows users to specify the current names of the columns that correspond to the columns required by scoringutils (observed, predicted, model, as well quantile_level for quantile-based forecasts and sample_id for sample-based forecasts). as_forecast() renames the existing columns.

allows users to specify the unit of a single forecast. It removes all columns that are neither part of the forecast unit nor a required column (see set_forecast_unit() for details)

Determines the forecast type (binary, point, sample-based or quantile-based) from the input data (using the function get_forecast_type().

Constructs a forecast object of the appropriate class (forecast_binary, forecast_point, forecast_sample, or forecast_quantile, using the function new_forecast()).

Calls validate_forecast() on the newly created forecast object to validate it

Usage

as_forecast(data, ...)

# S3 method for default
as_forecast(
  data,
  forecast_unit = NULL,
  forecast_type = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  quantile_level = NULL,
  sample_id = NULL,
  ...
)

Arguments

data

A data.frame (or similar) with predicted and observed values. See as_forecast() for additional information on input formats.

...

additional arguments

forecast_unit

(optional) Name of the columns in data (after any renaming of columns done by as_forecast()) that denote the unit of a single forecast. See get_forecast_unit() for details. If NULL (the default), all columns that are not required columns are assumed to form the unit of a single forecast. If specified, all columns that are not part of the forecast unit (or required columns) will be removed.

forecast_type

(optional) The forecast type you expect the forecasts to have. If the forecast type as determined by scoringutils based on the input does not match this, an error will be thrown. If NULL (the default), the forecast type will be inferred from the data.

observed

(optional) Name of the column in data that contains the observed values. This column will be renamed to "observed".

predicted

(optional) Name of the column in data that contains the predicted values. This column will be renamed to "predicted".

model

(optional) Name of the column in data that contains the names of the models/forecasters that generated the predicted values. This column will be renamed to "model".

quantile_level

(optional) Name of the column in data that contains the quantile level of the predicted values. This column will be renamed to "quantile_level". Only applicable to quantile-based forecasts.

sample_id

(optional) Name of the column in data that contains the sample id. This column will be renamed to "sample_id". Only applicable to sample-based forecasts.

Value

Depending on the forecast type, an object of class forecast_binary, forecast_point, forecast_sample or forecast_quantile.

Forecast types and input format

Various different forecast types / forecast formats are supported. At the moment, those are

point forecasts

binary forecasts ("soft binary classification")

Probabilistic forecasts in a quantile-based format (a forecast is represented as a set of predictive quantiles)

Probabilistic forecasts in a sample-based format (a forecast is represented as a set of predictive samples)

Forecast types are determined based on the columns present in the input data.

Point forecasts require a column observed of type numeric and a column predicted of type numeric.

Binary forecasts require a column observed of type factor with exactly two levels and a column predicted of type numeric with probabilities, corresponding to the probability that observed is equal to the second factor level. See details here for more information.

Quantile-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column quantile_level of type numeric with quantile-levels (between 0 and 1).

Sample-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column sample_id of type numeric with sample indices.

For more information see the vignettes and the example data (example_quantile, example_continuous, example_integer, example_point(), and example_binary).

Examples

as_forecast(example_binary)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> Forecast type:
#> binary
#> Forecast unit:
#> location, location_name, target_end_date, target_type, forecast_date, model,
#> and horizon
#> 
#>       location location_name target_end_date target_type forecast_date
#>         <char>        <char>          <Date>      <char>        <Date>
#>    1:       DE       Germany      2021-01-02       Cases          <NA>
#>    2:       DE       Germany      2021-01-02      Deaths          <NA>
#>    3:       DE       Germany      2021-01-09       Cases          <NA>
#>    4:       DE       Germany      2021-01-09      Deaths          <NA>
#>    5:       DE       Germany      2021-01-16       Cases          <NA>
#>   ---                                                                 
#> 1027:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 1028:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 1029:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 1030:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 1031:       IT         Italy      2021-07-24      Deaths    2021-07-12
#>                       model horizon predicted observed
#>                      <char>   <num>     <num>   <fctr>
#>    1:                  <NA>      NA        NA     <NA>
#>    2:                  <NA>      NA        NA     <NA>
#>    3:                  <NA>      NA        NA     <NA>
#>    4:                  <NA>      NA        NA     <NA>
#>    5:                  <NA>      NA        NA     <NA>
#>   ---                                                 
#> 1027: EuroCOVIDhub-baseline       2     0.250        0
#> 1028:       UMass-MechBayes       3     0.475        0
#> 1029:       UMass-MechBayes       2     0.450        0
#> 1030:  epiforecasts-EpiNow2       3     0.375        0
#> 1031:  epiforecasts-EpiNow2       2     0.300        0
as_forecast(
  example_quantile,
  forecast_unit = c("model", "target_type", "target_end_date",
                    "horizon", "location")
)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> Forecast type:
#> quantile
#> Forecast unit:
#> model, target_type, target_end_date, horizon, and location
#> 
#> Key: <location, target_end_date, target_type>
#>        observed quantile_level predicted                model target_type
#>           <num>          <num>     <int>               <char>      <char>
#>     1:   127300             NA        NA                 <NA>       Cases
#>     2:     4534             NA        NA                 <NA>      Deaths
#>     3:   154922             NA        NA                 <NA>       Cases
#>     4:     6117             NA        NA                 <NA>      Deaths
#>     5:   110183             NA        NA                 <NA>       Cases
#>    ---                                                                   
#> 20541:       78          0.850       352 epiforecasts-EpiNow2      Deaths
#> 20542:       78          0.900       397 epiforecasts-EpiNow2      Deaths
#> 20543:       78          0.950       499 epiforecasts-EpiNow2      Deaths
#> 20544:       78          0.975       611 epiforecasts-EpiNow2      Deaths
#> 20545:       78          0.990       719 epiforecasts-EpiNow2      Deaths
#>        target_end_date horizon location
#>                 <Date>   <num>   <char>
#>     1:      2021-01-02      NA       DE
#>     2:      2021-01-02      NA       DE
#>     3:      2021-01-09      NA       DE
#>     4:      2021-01-09      NA       DE
#>     5:      2021-01-16      NA       DE
#>    ---                                 
#> 20541:      2021-07-24       2       IT
#> 20542:      2021-07-24       2       IT
#> 20543:      2021-07-24       2       IT
#> 20544:      2021-07-24       2       IT
#> 20545:      2021-07-24       2       IT