Determines bias of forecasts

Determines bias from predictive Monte-Carlo samples. The function automatically recognises, whether forecasts are continuous or integer valued and adapts the Bias function accordingly.

bias(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector of length n with the biases of the predictive samples with respect to the true values.

Details

For continuous forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - 2 * (P_t (x_t)) $$

where \(P_t\) is the empirical cumulative distribution function of the prediction for the true value \(x_t\). Computationally, \(P_t (x_t)\) is just calculated as the fraction of predictive samples for \(x_t\) that are smaller than \(x_t\).

For integer valued forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - (P_t (x_t) + P_t (x_t + 1)) $$

to adjust for the integer nature of the forecasts.

In both cases, Bias can assume values between -1 and 1 and is 0 ideally.

References

The integer valued Bias function is discussed in Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15 Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, et al. (2019) Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15. PLOS Computational Biology 15(2): e1006785. https://doi.org/10.1371/journal.pcbi.1006785

Author

Nikos Bosse nikosbosse@gmail.com

Examples

## integer valued forecasts
true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
bias(true_values, predictions)
#>  [1]  0.635  0.845  0.725  0.710 -0.785 -0.060 -0.410 -0.090  0.180  0.070
#> [11] -0.845  0.220 -0.250  0.195 -0.715 -0.470 -0.225 -0.245 -0.530 -0.745
#> [21] -0.815  0.355  0.970  0.980  0.540 -0.520  0.425 -0.680  0.515  0.500

## continuous forecasts
true_values <- rnorm(30, mean = 1:30)
predictions <- replicate(200, rnorm(30, mean = 1:30))
bias(true_values, predictions)
#>  [1]  0.84 -0.03  0.33 -0.38  0.59 -0.40 -0.60  0.96 -0.61 -0.04 -0.79 -0.77
#> [13]  0.67 -0.30  0.10  0.39  0.92  0.72 -0.90  0.98  0.73  0.01  0.39 -0.22
#> [25]  0.20  0.07 -0.63  0.58 -0.32  0.31