Brier Score

Computes the Brier Score for probabilistic forecasts of binary outcomes.

Usage

brier_score(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n with all values equal to either 0 or 1

predictions

A vector with a predicted probability that true_value = 1.

Value

A numeric value with the Brier Score, i.e. the mean squared error of the given probability forecasts

Details

The Brier score is a proper score rule that assesses the accuracy of probabilistic binary predictions. The outcomes can be either 0 or 1, the predictions must be a probability that the true outcome will be 1.

The Brier Score is then computed as the mean squared error between the probabilistic prediction and the true outcome.

$$ \textrm{Brier\_Score} = \frac{1}{N} \sum_{t = 1}^{n} (\textrm{prediction}_t - \textrm{outcome}_t)^2 $$

Examples

true_values <- sample(c(0, 1), size = 30, replace = TRUE)
predictions <- runif(n = 30, min = 0, max = 1)

brier_score(true_values, predictions)
#>  [1] 0.001421642 0.188324691 0.020988001 0.007578324 0.220852390 0.611838777
#>  [7] 0.722062587 0.526106217 0.234673636 0.242647525 0.206171819 0.202997979
#> [13] 0.164195697 0.047030607 0.007113980 0.015254515 0.314339635 0.176935113
#> [19] 0.662900669 0.646656534 0.005554719 0.817252294 0.127739610 0.138462649
#> [25] 0.294423747 0.065351374 0.017987251 0.788861055 0.411043427 0.616810147