Check forecasts

Function to check the input data before running eval_forecasts().

The data should come in one of three different formats:

A format for binary predictions (see binary_example_data)

A sample-based format for discrete or continuous predictions (see continuous_example_data and integer_example_data)

A quantile-based format (see quantile_example_data)

check_forecasts(data)

Arguments

data

A data.frame or similar as would be used for eval_forecasts()

Value

A list with elements that give information about what scoringutilsthinks you are trying to do and potential issues.

target_type the type of the prediction target as inferred from the input: 'binary', if all values in true_value are either 0 or 1 and values in prediction are between 0 and 1, 'discrete' if all true values are integers and 'continuous' if not.

prediction_type inferred type of the prediction. 'quantile', if there is a column called 'quantile', else 'discrete' if all values in prediction are integer, else 'continuous.

forecast_unit unit of a single forecast, i.e. the grouping that uniquely defines a single forecast. This is assumed to be all present columns apart from the following protected columns: c("prediction", "true_value", "sample", "quantile","range", "boundary"). It is important that you remove all unnecessary columns before scoring.

rows_per_forecast a data.frame that shows how many rows (usually quantiles or samples there are available per forecast. If a forecast model has several entries, then there a forecasts with differing numbers of quantiles / samples.

unique_values A data.frame that shows how many unique values there are present per model and column in the data. This doesn't directly show missing values, but rather the maximum number of unique values across the whole data.

warnings A vector with warnings. These can be ignored if you know what you are doing.

errors A vector with issues that will cause an error when running eval_forecasts().

messages A verbal explanation of the information provided above.

See also

Author

Nikos Bosse nikosbosse@gmail.com

Examples

library(scoringutils)
check <- check_forecasts(quantile_example_data)
print(check)
#> $target_type
#> [1] "integer"
#> 
#> $prediction_type
#> [1] "quantile"
#> 
#> $forecast_unit
#> [1] "geography"     "value_type"    "value_desc"    "value_date"   
#> [5] "model"         "creation_date" "horizon"      
#> 
#> $rows_per_forecast
#>                 model rows_per_forecast
#> 1:            EpiSoon                19
#> 2: DetSEIRwithNB MCMC                19
#> 3:           SIRCOVID                19
#> 
#> $unique_values
#>                  model  V1
#>  1:            EpiSoon   4
#>  2:            EpiSoon   2
#>  3:            EpiSoon   2
#>  4:            EpiSoon   8
#>  5:            EpiSoon  16
#>  6:            EpiSoon   6
#>  7:            EpiSoon  19
#>  8:            EpiSoon 244
#>  9:            EpiSoon   3
#> 10: DetSEIRwithNB MCMC   4
#> 11: DetSEIRwithNB MCMC   3
#> 12: DetSEIRwithNB MCMC   3
#> 13: DetSEIRwithNB MCMC   8
#> 14: DetSEIRwithNB MCMC  34
#> 15: DetSEIRwithNB MCMC   6
#> 16: DetSEIRwithNB MCMC  19
#> 17: DetSEIRwithNB MCMC 516
#> 18: DetSEIRwithNB MCMC   3
#> 19:           SIRCOVID   4
#> 20:           SIRCOVID   2
#> 21:           SIRCOVID   2
#> 22:           SIRCOVID   8
#> 23:           SIRCOVID  25
#> 24:           SIRCOVID   6
#> 25:           SIRCOVID  19
#> 26:           SIRCOVID 399
#> 27:           SIRCOVID   3
#>                  model  V1
#> 
#> $warnings
#> [1] "Some values for `prediction` are NA in the data provided"
#> 
#> 
#> Based on your input, scoringutils thinks:
#> Forecasts are for a `integer` target using a `quantile` prediction format.
#> The unit of a single forecast is defined by `geography`, `value_type`, `value_desc`, `value_date`, `model`, `creation_date`, `horizon`. If this is not as intended, please DELETE UNNECESSARY columns or add new ones.
#> $rows_per_forecast shows how many rows (usually quantiles or samples are available per forecast.
#> $unique_values shows how many unique values there are per column per model (across the entire data).
#> 
#> You should be aware of the following warnings:
#> Some values for `prediction` are NA in the data provided
check_forecasts(binary_example_data)
#> $target_type
#> [1] "binary"
#> 
#> $prediction_type
#> [1] "continuous"
#> 
#> $forecast_unit
#> [1] "value_date"    "value_type"    "geography"     "value_desc"   
#> [5] "model"         "creation_date" "horizon"      
#> 
#> $rows_per_forecast
#>                 model rows_per_forecast
#> 1:            EpiSoon                 1
#> 2: DetSEIRwithNB MCMC                 1
#> 3:           SIRCOVID                 1
#> 
#> $unique_values
#>                  model V1
#>  1:            EpiSoon  8
#>  2:            EpiSoon  2
#>  3:            EpiSoon  4
#>  4:            EpiSoon  2
#>  5:            EpiSoon  6
#>  6:            EpiSoon  3
#>  7:            EpiSoon 21
#>  8:            EpiSoon  2
#>  9: DetSEIRwithNB MCMC  8
#> 10: DetSEIRwithNB MCMC  4
#> 11: DetSEIRwithNB MCMC  4
#> 12: DetSEIRwithNB MCMC  4
#> 13: DetSEIRwithNB MCMC  6
#> 14: DetSEIRwithNB MCMC  3
#> 15: DetSEIRwithNB MCMC 24
#> 16: DetSEIRwithNB MCMC  2
#> 17:           SIRCOVID  8
#> 18:           SIRCOVID  3
#> 19:           SIRCOVID  4
#> 20:           SIRCOVID  3
#> 21:           SIRCOVID  6
#> 22:           SIRCOVID  3
#> 23:           SIRCOVID 18
#> 24:           SIRCOVID  2
#>                  model V1
#> 
#> $warnings
#> [1] "Some values for `true_value` are NA in the data provided"
#> 
#> 
#> Based on your input, scoringutils thinks:
#> Forecasts are for a `binary` target using a `continuous` prediction format.
#> The unit of a single forecast is defined by `value_date`, `value_type`, `geography`, `value_desc`, `model`, `creation_date`, `horizon`. If this is not as intended, please DELETE UNNECESSARY columns or add new ones.
#> $rows_per_forecast shows how many rows (usually quantiles or samples are available per forecast.
#> $unique_values shows how many unique values there are per column per model (across the entire data).
#> 
#> You should be aware of the following warnings:
#> Some values for `true_value` are NA in the data provided