Ranked Probability Score

Wrapper around the crps_sample() function from the scoringRules package. Can be used for continuous as well as integer valued forecasts

crps(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with the scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts with scoringRules, https://arxiv.org/pdf/1709.04743.pdf

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
crps(true_values, predictions)
#>  [1] 0.528475 0.526575 0.571975 1.190125 1.898600 0.680450 1.620725 1.610200
#>  [9] 0.758950 2.866175 2.292125 0.777825 3.003475 2.789375 0.838600 1.112850
#> [17] 1.041450 4.318375 1.069575 3.194525 2.006425 4.757575 2.242850 2.937975
#> [25] 1.357550 3.230275 5.649800 3.577125 4.534150 6.126275