Ranked Probability Score

Wrapper around the crps_sample() function from the scoringRules package. Can be used for continuous as well as integer valued forecasts

Usage

crps_sample(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n.

Value

vector with the scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts with scoringRules, https://www.jstatsoft.org/article/view/v090i12

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
crps_sample(true_values, predictions)
#>  [1] 0.394375 0.652725 1.230525 1.844400 1.331500 3.837250 0.517825 0.747850
#>  [9] 1.550400 1.706400 1.723675 5.628025 0.873250 1.195400 3.914475 3.037525
#> [17] 1.143400 0.880650 3.472625 1.331925 1.855375 1.830725 1.494125 1.039725
#> [25] 1.345700 1.393375 1.651300 4.729075 3.292425 1.270025