Dawid-Sebastiani Score

Wrapper around the dss_sample function from the scoringRules package.

dss(true_values, predictions)

Arguments

true_values
A vector with the true observed values of size n

predictions
nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with scoring values

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
dss(true_values, predictions)

#>  [1] 0.8769948 0.5158188 1.4255518 6.0934775 6.5835359 2.0063774 2.6685313
#>  [8] 4.1364559 2.4630771 4.1437293 2.4178409 5.8180589 3.4426199 3.2187074
#> [15] 4.0637667 4.2718337 3.3061971 3.1069574 2.9170516 3.6404485 3.4208558
#> [22] 3.1639120 3.4272242 3.8340961 8.9398436 3.4818540 3.3227001 5.3394811
#> [29] 6.0862997 3.4147784