Dawid-Sebastiani Score

Wrapper around the dss_sample() function from the scoringRules package.

Usage

dss_sample(observed, predicted, ...)

Arguments

observed

A vector with observed values of size n

predicted

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predicted can just be a vector of size n.

...

additional arguments passed to dss_sample() from the scoringRules package.

Value

vector with scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts with scoringRules, https://www.jstatsoft.org/article/view/v090i12

Examples

observed <- rpois(30, lambda = 1:30)
predicted <- replicate(200, rpois(n = 30, lambda = 1:30))
dss_sample(observed, predicted)
#>  [1] -0.06713812  1.17180200  1.42079340  2.12277933  3.37480157  4.38080715
#>  [7]  2.13959988  3.04514163  2.44538589  3.32679936  5.75993445  4.27538513
#> [13]  3.46558702  2.86525635  4.63714888  2.78870812  3.09324541  4.36007539
#> [19]  3.22285135  3.30833409  3.67328596  5.87385691  3.26446949  4.29307072
#> [25]  4.22441937  3.43044643  3.44780381  3.41887494  3.40691223  4.53403296