Evaluate forecasts

The function eval_forecasts is a wrapper that provides an interface to lower-level functions. It can be used to assess the goodness of probabilistic or point forecasts to continues, integer-valued or binary variables. The lower-level functions accessed are:

eval_forecasts_prob_int

eval_forecasts_prob_cont

eval_forecasts_prob_bin

eval_forecasts_point_int

eval_forecasts_point_cont

eval_forecasts_point_bin

eval_forecasts(
  true_values,
  predictions,
  prediction_type = "probabilistic",
  outcome_type = "integer",
  metrics = NULL,
  output = "df"
)

Arguments

true_values
A vector with the true observed values of size n

predictions
a list of appropriate predictions. Every item in the list corresponds to the predictions made by one model. Appropriate dimensions and input types are:

for probabilistic integer and continuous forecasts: a matrix or data.frame of size nxN of predictive samples. n (number of rows) being the number of observed values to predict and N (number of columns) the number of Monte Carlo samples

for probabilistic binary forecasts: a vector of length n that gives the probability that the corresponding element in true_values will be equal to one.

for all point estimates: a vector of size n with predictions for the corresponding entries of true_values

prediction_type
Type of the prediction made. Can be eitehr "probabilitic" or "point" for point predictions.

outcome_type
type of the variable to be predicted. Can be either "integer" or "continuous" or "binary".

metrics
what metrics to include. Currently not used, as all metrics are displayed

output
specify the format of the output. Can be either "df" (returns a single data.frame) or anything else (returns a list of data.frames)

Value

output option "df" returns a single data.frame with the prediction results. Rownames of the data.frame correspond to the metric applied for the scoring. mean and sd are the mean and standard deviations of the scores achieved by the predictions for every single value of true_values. Only in the case of the pit, mean and sd return the mean and standard deviation of the Replicates of the Randomised PIT. If everything else than "df" is specified, the above results are returned as a list of data.frames for the different metrics.

References

Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, Edmunds WJ (2019) Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15. PLoS Comput Biol 15(2): e1006785. https://doi.org/10.1371/journal.pcbi.1006785