Quantile example data

A data set with predictions for COVID-19 cases and deaths submitted to the European Forecast Hub.

Usage

example_quantile

Format

An object of class forecast_quantile (see as_forecast_quantile()) with the following columns:

location

the country for which a prediction was made

target_end_date

the date for which a prediction was made

target_type

the target to be predicted (cases or deaths)

observed

Numeric: observed values

location_name

name of the country for which a prediction was made

forecast_date

the date on which a prediction was made

quantile_level

quantile level of the corresponding prediction

predicted

predicted value

model

name of the model that generated the forecasts

horizon

forecast horizon in weeks

Source

https://github.com/european-modelling-hubs/covid19-forecast-hub-europe_archive/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/

Details

The data was created using the script create-example-data.R in the inst/ folder (or the top level folder in a compiled package).