Discrete forecast example data

A data set with integer predictions for COVID-19 cases and deaths constructed from data submitted to the European Forecast Hub.

Usage

example_sample_discrete

Format

A data frame with the following columns:

location

the country for which a prediction was made

target_end_date

the date for which a prediction was made

target_type

the target to be predicted (cases or deaths)

observed

observed values

location_name

name of the country for which a prediction was made

forecast_date

the date on which a prediction was made

model

name of the model that generated the forecasts

horizon

forecast horizon in weeks

predicted

predicted value

sample_id

id for the corresponding sample

Source

https://github.com/european-modelling-hubs/covid19-forecast-hub-europe/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/

Details

The data was created using the script create-example-data.R in the inst/ folder (or the top level folder in a compiled package).