Get metrics

Generic function to to obtain default metrics available for scoring or metrics that were used for scoring.

If called on a forecast object it returns a list of functions that can be used for scoring.

If called on a scores object (see score()), it returns a character vector with the names of the metrics that were used for scoring.

See the documentation for the actual methods in the See Also section below for more details. Alternatively call ?get_metrics.<forecast_type> or ?get_metrics.scores.

Usage

get_metrics(x, ...)

Arguments

x

A forecast or scores object.

...

Additional arguments passed to the method.

See also