PIT Histogram

Make a simple histogram of the probability integral transformed values to visually check whether a uniform distribution seems likely.

hist_PIT(PIT_samples, num_bins = NULL)

Arguments

PIT_samples
A vector with the PIT values of size n

num_bins
the number of bins in the PIT histogram. If not given, the square root of n will be used

Value

vector with the scoring values

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
logs(true_values, predictions)

#>  [1] 1.731332 1.209836 1.968243 1.507185 2.228183 2.113638 2.951221 3.094222
#>  [9] 2.051460 4.196550 3.322451 2.309225 3.580306 2.234037 2.636911 2.787277
#> [17] 2.737110 2.453792 2.543680 3.922050 2.587081 2.953560 2.516912 3.022183
#> [25] 2.690586 4.940312 3.131906 2.935527 4.191581 3.006766