Integer Example Data

A toy dataset for a probabilistic forecast of an integer outcome variable

integer_example_data

Format

A data frame with 6000 rows and 5 variables:

id

unique identifier for true observed values

model

name of the model that generated the forecasts

horizon

forecast horzin (e.g. 1 day ahead forecast)

true_value

true observed values

sample

number that identifies the predictive sample generated by a specific model for a specific observed value

prediction

predictive sample for the corresponding true value