LogS

Wrapper around the logs_sample function from the scoringRules package. While the Log Score is in theory also applicable to integer forecasts, the problem lies in the implementation: The Log Score needs a kernel density estimation, which is not well defined with integer-valued Monte Carlo Samples. The Log Score can be used for specific integer valued probabiliy distributions. See the scoringRules package for more details.

logs(true_values, predictions)

Arguments

true_values
A vector with the true observed values of size n

predictions
nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with the scoring values

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
logs(true_values, predictions)

#>  [1] 1.515553 3.185156 1.842941 3.269126 2.434908 2.066662 2.083242 1.965935
#>  [9] 2.687376 3.874522 2.225258 2.195834 2.267778 3.299158 3.151222 2.955654
#> [17] 4.499432 2.973252 2.474821 2.756225 3.449377 3.541374 2.762083 3.369988
#> [25] 2.578536 3.082390 2.890874 3.301030 2.827223 2.813937