Scoring rules for nominal forecasts

Helper function that returns a named list of default scoring rules suitable for nominal forecasts.

The default scoring rules are:

"log_score" = logs_nominal()

Usage

metrics_nominal(select = NULL, exclude = NULL)

Arguments

select

A character vector of scoring rules to select from the list. If select is NULL (the default), all possible scoring rules are returned.

exclude

A character vector of scoring rules to exclude from the list. If select is not NULL, this argument is ignored.

Value

A list of scoring functions.

Examples

metrics_nominal()
#> $log_score
#> function (observed, predicted, predicted_label) 
#> {
#>     assert_input_nominal(observed, predicted, predicted_label)
#>     n <- length(observed)
#>     if (n == 1) {
#>         predicted <- matrix(predicted, nrow = 1)
#>     }
#>     observed_indices <- as.numeric(observed)
#>     pred_for_observed <- predicted[cbind(1:n, observed_indices)]
#>     logs <- -log(pred_for_observed)
#>     return(logs)
#> }
#> <bytecode: 0x55c9348a0448>
#> <environment: namespace:scoringutils>
#> 
metrics_nominal(select = "log_score")
#> $log_score
#> function (observed, predicted, predicted_label) 
#> {
#>     assert_input_nominal(observed, predicted, predicted_label)
#>     n <- length(observed)
#>     if (n == 1) {
#>         predicted <- matrix(predicted, nrow = 1)
#>     }
#>     observed_indices <- as.numeric(observed)
#>     pred_for_observed <- predicted[cbind(1:n, observed_indices)]
#>     logs <- -log(pred_for_observed)
#>     return(logs)
#> }
#> <bytecode: 0x55c9348a0448>
#> <environment: namespace:scoringutils>
#>