Pivot Range Format Forecasts From Long to Wide Format

Given a data.frame that follows the structure shown in range_example_data_long, the function outputs the same data in a long format as (as shown in range_example_data_wide). This can be useful e.g. for plotting.

range_long_to_wide(data)

Arguments

data
a data.frame following the specifications from eval_forecasts) for quantile forecasts. For an example, see range_example_data_long)

Value

a data.frame in wide format

Examples

long <- scoringutils::range_example_data_long
wide <- scoringutils::range_long_to_wide(long)