Scoring Rules for Quantile-Based Forecasts

Helper function that returns a named list of default scoring rules suitable for forecasts in a quantile-based format

The default scoring rules are:

"wis" = wis

"overprediction" = overprediction()

"underprediction" = underprediction()

"dispersion" = dispersion()

"bias" = bias_quantile()

"interval_coverage_50" = interval_coverage()

"interval_coverage_90" = function(...) { run_safely(..., interval_range = 90, fun = interval_coverage) }

"interval_coverage_deviation" = interval_coverage_deviation(),

"ae_median" = ae_median_quantile()

Note: The interval_coverage_90 scoring rule is created as a wrapper around interval_coverage(), making use of the function run_safely(). This construct allows the function to deal with arbitrary arguments in ..., while making sure that only those that interval_coverage() can accept get passed on to it. interval_range = 90 is set in the function definition, as passing an argument interval_range = 90 to score() would mean it would also get passed to interval_coverage_50.

Usage

rules_quantile(select = NULL, exclude = NULL)

Arguments

select

A character vector of scoring rules to select from the list. If select is NULL (the default), all possible scoring rules are returned.

exclude

A character vector of scoring rules to exclude from the list. If select is not NULL, this argument is ignored.

Value

A list of scoring rules.

Examples

rules_quantile()
#> $wis
#> function (observed, predicted, quantile_level, separate_results = FALSE, 
#>     weigh = TRUE, count_median_twice = FALSE, na.rm = TRUE) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     reformatted <- quantile_to_interval(observed, predicted, 
#>         quantile_level)
#>     assert_logical(separate_results, len = 1)
#>     assert_logical(weigh, len = 1)
#>     assert_logical(count_median_twice, len = 1)
#>     assert_logical(na.rm, len = 1)
#>     if (separate_results) {
#>         cols <- c("wis", "dispersion", "underprediction", "overprediction")
#>     }
#>     else {
#>         cols <- "wis"
#>     }
#>     reformatted[, `:=`(eval(cols), do.call(interval_score, list(observed = observed, 
#>         lower = lower, upper = upper, interval_range = interval_range, 
#>         weigh = weigh, separate_results = separate_results)))]
#>     if (count_median_twice) {
#>         reformatted[, `:=`(weight, 1)]
#>     }
#>     else {
#>         reformatted[, `:=`(weight, ifelse(interval_range == 0, 
#>             0.5, 1))]
#>     }
#>     reformatted <- reformatted[, lapply(.SD, weighted.mean, na.rm = na.rm, 
#>         w = weight), by = "forecast_id", .SDcols = colnames(reformatted) %like% 
#>         paste(cols, collapse = "|")]
#>     if (separate_results) {
#>         return(list(wis = reformatted$wis, dispersion = reformatted$dispersion, 
#>             underprediction = reformatted$underprediction, overprediction = reformatted$overprediction))
#>     }
#>     else {
#>         return(reformatted$wis)
#>     }
#> }
#> <bytecode: 0x55a1c9024da0>
#> <environment: namespace:scoringutils>
#> 
#> $overprediction
#> function (observed, predicted, quantile_level, ...) 
#> {
#>     args <- list(...)
#>     args$separate_results <- TRUE
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     out <- do.call(wis, c(list(observed), list(predicted), list(quantile_level), 
#>         args))
#>     return(out$overprediction)
#> }
#> <bytecode: 0x55a1c9036eb8>
#> <environment: namespace:scoringutils>
#> 
#> $underprediction
#> function (observed, predicted, quantile_level, ...) 
#> {
#>     args <- list(...)
#>     args$separate_results <- TRUE
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     out <- do.call(wis, c(list(observed), list(predicted), list(quantile_level), 
#>         args))
#>     return(out$underprediction)
#> }
#> <bytecode: 0x55a1c903b890>
#> <environment: namespace:scoringutils>
#> 
#> $dispersion
#> function (observed, predicted, quantile_level, ...) 
#> {
#>     args <- list(...)
#>     args$separate_results <- TRUE
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     out <- do.call(wis, c(list(observed), list(predicted), list(quantile_level), 
#>         args))
#>     return(out$dispersion)
#> }
#> <bytecode: 0x55a1c903a550>
#> <environment: namespace:scoringutils>
#> 
#> $bias
#> function (observed, predicted, quantile_level, na.rm = TRUE) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     n <- length(observed)
#>     N <- length(quantile_level)
#>     if (is.null(dim(predicted))) {
#>         dim(predicted) <- c(n, N)
#>     }
#>     if (!(0.5 %in% quantile_level)) {
#>         cli_inform(c(i = "Median not available, computing bias as mean of the two\n        innermost quantiles in order to compute bias."))
#>     }
#>     bias <- sapply(1:n, function(i) {
#>         bias_quantile_single_vector(observed[i], predicted[i, 
#>             ], quantile_level, na.rm)
#>     })
#>     return(bias)
#> }
#> <bytecode: 0x55a1c9a36a70>
#> <environment: namespace:scoringutils>
#> 
#> $interval_coverage_50
#> function (observed, predicted, quantile_level, interval_range = 50) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     assert_number(interval_range)
#>     necessary_quantiles <- c((100 - interval_range)/2, 100 - 
#>         (100 - interval_range)/2)/100
#>     if (!all(necessary_quantiles %in% quantile_level)) {
#>         cli_warn(c(`!` = "To compute the interval coverage for an interval range of\n        {.val {interval_range}%}, the {.val {necessary_quantiles}} quantiles\n        are required.", 
#>             i = "Returning {.val {NA}}."))
#>         return(NA)
#>     }
#>     r <- interval_range
#>     reformatted <- quantile_to_interval(observed, predicted, 
#>         quantile_level)
#>     reformatted <- reformatted[interval_range %in% r]
#>     reformatted[, `:=`(interval_coverage, (observed >= lower) & 
#>         (observed <= upper))]
#>     return(reformatted$interval_coverage)
#> }
#> <bytecode: 0x55a1c903ee10>
#> <environment: namespace:scoringutils>
#> 
#> $interval_coverage_90
#> function (...) 
#> {
#>     run_safely(..., interval_range = 90, fun = interval_coverage)
#> }
#> <bytecode: 0x55a1c9025890>
#> <environment: 0x55a1cb2a9f48>
#> 
#> $interval_coverage_deviation
#> function (observed, predicted, quantile_level) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     available_ranges <- unique(get_range_from_quantile(quantile_level))
#>     necessary_quantiles <- unique(c((100 - available_ranges)/2, 
#>         100 - (100 - available_ranges)/2)/100)
#>     if (!all(necessary_quantiles %in% quantile_level)) {
#>         missing <- necessary_quantiles[!necessary_quantiles %in% 
#>             quantile_level]
#>         cli_warn(c(x = "To compute interval coverage deviation, all quantiles must form\n        central symmetric prediction intervals.", 
#>             i = "Missing quantiles: {.val {missing}}. Returning {.val {NA}}."))
#>         return(NA)
#>     }
#>     reformatted <- quantile_to_interval(observed, predicted, 
#>         quantile_level)[interval_range != 0]
#>     reformatted[, `:=`(interval_coverage, (observed >= lower) & 
#>         (observed <= upper))]
#>     reformatted[, `:=`(interval_coverage_deviation, interval_coverage - 
#>         interval_range/100)]
#>     out <- reformatted[, .(interval_coverage_deviation = mean(interval_coverage_deviation)), 
#>         by = "forecast_id"]
#>     return(out$interval_coverage_deviation)
#> }
#> <bytecode: 0x55a1c90445c0>
#> <environment: namespace:scoringutils>
#> 
#> $ae_median
#> function (observed, predicted, quantile_level) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     if (!any(quantile_level == 0.5)) {
#>         cli_warn(c(x = "In order to compute the absolute error of the median,\n        {.val 0.5} must be among the quantiles given.", 
#>             i = "Returning {.val NA}."))
#>         return(NA_real_)
#>     }
#>     if (is.null(dim(predicted))) {
#>         predicted <- matrix(predicted, nrow = 1)
#>     }
#>     predicted <- predicted[, quantile_level == 0.5]
#>     abs_error_median <- abs(observed - predicted)
#>     return(abs_error_median)
#> }
#> <bytecode: 0x55a1c9333880>
#> <environment: namespace:scoringutils>
#> 
rules_quantile(select = "wis")
#> $wis
#> function (observed, predicted, quantile_level, separate_results = FALSE, 
#>     weigh = TRUE, count_median_twice = FALSE, na.rm = TRUE) 
#> {
#>     assert_input_quantile(observed, predicted, quantile_level)
#>     reformatted <- quantile_to_interval(observed, predicted, 
#>         quantile_level)
#>     assert_logical(separate_results, len = 1)
#>     assert_logical(weigh, len = 1)
#>     assert_logical(count_median_twice, len = 1)
#>     assert_logical(na.rm, len = 1)
#>     if (separate_results) {
#>         cols <- c("wis", "dispersion", "underprediction", "overprediction")
#>     }
#>     else {
#>         cols <- "wis"
#>     }
#>     reformatted[, `:=`(eval(cols), do.call(interval_score, list(observed = observed, 
#>         lower = lower, upper = upper, interval_range = interval_range, 
#>         weigh = weigh, separate_results = separate_results)))]
#>     if (count_median_twice) {
#>         reformatted[, `:=`(weight, 1)]
#>     }
#>     else {
#>         reformatted[, `:=`(weight, ifelse(interval_range == 0, 
#>             0.5, 1))]
#>     }
#>     reformatted <- reformatted[, lapply(.SD, weighted.mean, na.rm = na.rm, 
#>         w = weight), by = "forecast_id", .SDcols = colnames(reformatted) %like% 
#>         paste(cols, collapse = "|")]
#>     if (separate_results) {
#>         return(list(wis = reformatted$wis, dispersion = reformatted$dispersion, 
#>             underprediction = reformatted$underprediction, overprediction = reformatted$overprediction))
#>     }
#>     else {
#>         return(reformatted$wis)
#>     }
#> }
#> <bytecode: 0x55a1c9024da0>
#> <environment: namespace:scoringutils>
#>