Scoring Rules for Sample-Based Forecasts

Helper function that returns a named list of default scoring rules suitable for forecasts in a sample-based format

The default scoring rules are:

"mad" = mad_sample()

"bias" = bias_sample()

"dss" = dss_sample()

"crps" = crps_sample()

"log_score" = logs_sample()

"mad" = mad_sample()

"ae_median" = ae_median_sample()

"se_mean" = se_mean_sample()

Usage

rules_sample(select = NULL, exclude = NULL)

Arguments

select

A character vector of scoring rules to select from the list. If select is NULL (the default), all possible scoring rules are returned.

exclude

A character vector of scoring rules to exclude from the list. If select is not NULL, this argument is ignored.

Value

A list of scoring rules.

Examples

rules_sample()
#> $bias
#> function (observed, predicted) 
#> {
#>     assert_input_sample(observed, predicted)
#>     prediction_type <- get_type(predicted)
#>     n_pred <- ncol(predicted)
#>     p_x <- rowSums(predicted <= observed)/n_pred
#>     if (prediction_type == "continuous") {
#>         res <- 1 - 2 * p_x
#>         return(res)
#>     }
#>     else {
#>         p_xm1 <- rowSums(predicted <= (observed - 1))/n_pred
#>         res <- 1 - (p_x + p_xm1)
#>         return(res)
#>     }
#> }
#> <bytecode: 0x56007e47a800>
#> <environment: namespace:scoringutils>
#> 
#> $dss
#> function (observed, predicted, ...) 
#> {
#>     assert_input_sample(observed, predicted)
#>     scoringRules::dss_sample(y = observed, dat = predicted, ...)
#> }
#> <bytecode: 0x560080e35030>
#> <environment: namespace:scoringutils>
#> 
#> $crps
#> function (observed, predicted, ...) 
#> {
#>     assert_input_sample(observed, predicted)
#>     scoringRules::crps_sample(y = observed, dat = predicted, 
#>         ...)
#> }
#> <bytecode: 0x56007f039260>
#> <environment: namespace:scoringutils>
#> 
#> $log_score
#> function (observed, predicted, ...) 
#> {
#>     assert_input_sample(observed, predicted)
#>     scoringRules::logs_sample(y = observed, dat = predicted, 
#>         ...)
#> }
#> <bytecode: 0x56007cf47da0>
#> <environment: namespace:scoringutils>
#> 
#> $mad
#> function (observed = NULL, predicted, ...) 
#> {
#>     assert_input_sample(rep(NA_real_, nrow(predicted)), predicted)
#>     sharpness <- apply(predicted, MARGIN = 1, mad, ...)
#>     return(sharpness)
#> }
#> <bytecode: 0x560082b8f708>
#> <environment: namespace:scoringutils>
#> 
#> $ae_median
#> function (observed, predicted) 
#> {
#>     assert_input_sample(observed, predicted)
#>     median_predictions <- apply(as.matrix(predicted), MARGIN = 1, 
#>         FUN = median)
#>     ae_median <- abs(observed - median_predictions)
#>     return(ae_median)
#> }
#> <bytecode: 0x56007f49d818>
#> <environment: namespace:scoringutils>
#> 
#> $se_mean
#> function (observed, predicted) 
#> {
#>     assert_input_sample(observed, predicted)
#>     mean_predictions <- rowMeans(as.matrix(predicted))
#>     se_mean <- (observed - mean_predictions)^2
#>     return(se_mean)
#> }
#> <bytecode: 0x5600822ac328>
#> <environment: namespace:scoringutils>
#> 
rules_sample(select = "mad")
#> $mad
#> function (observed = NULL, predicted, ...) 
#> {
#>     assert_input_sample(rep(NA_real_, nrow(predicted)), predicted)
#>     sharpness <- apply(predicted, MARGIN = 1, mad, ...)
#>     return(sharpness)
#> }
#> <bytecode: 0x560082b8f708>
#> <environment: namespace:scoringutils>
#>