Evaluate forecasts in a data.frame format

score() applies a selection of scoring metrics to a data.frame of forecasts. It is the workhorse of the scoringutils package. score() is a generic that dispatches to different methods depending on the class of the input data.

We recommend that users call as_forecast() prior to calling score() to validate the input data and convert it to a forecast object (though score.default() will do this if it hasn't happened before). See below for more information on forecast types and input formats. For additional help and examples, check out the Getting Started Vignette as well as the paper Evaluating Forecasts with scoringutils in R.

Usage

score(data, ...)

# S3 method for default
score(data, ...)

# S3 method for forecast_binary
score(data, metrics = metrics_binary, ...)

# S3 method for forecast_point
score(data, metrics = metrics_point, ...)

# S3 method for forecast_sample
score(data, metrics = metrics_sample, ...)

# S3 method for forecast_quantile
score(data, metrics = metrics_quantile, ...)

Arguments

data

A data.frame or data.table with predicted and observed values.

...

additional arguments

metrics

A named list of scoring functions. Names will be used as column names in the output. See metrics_point(), metrics_binary(), metrics_quantile(), and metrics_sample() for more information on the default metrics used.

Value

A data.table with unsummarised scores. This will generally be one score per forecast (as defined by the unit of a single forecast).

For quantile-based forecasts, one score per quantile will be returned instead. This is done as scores can be computed and may be of interest for individual quantiles. You can call summarise_scores()) on the unsummarised scores to obtain one score per forecast unit for quantile-based forecasts.

Forecast types and input format

Various different forecast types / forecast formats are supported. At the moment, those are

point forecasts

binary forecasts ("soft binary classification")

Probabilistic forecasts in a quantile-based format (a forecast is represented as a set of predictive quantiles)

Probabilistic forecasts in a sample-based format (a forecast is represented as a set of predictive samples)

Forecast types are determined based on the columns present in the input data.

Point forecasts require a column observed of type numeric and a column predicted of type numeric.

Binary forecasts require a column observed of type factor with exactly two levels and a column predicted of type numeric with probabilities, corresponding to the probability that observed is equal to the second factor level. See details here for more information.

Quantile-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column quantile of type numeric with quantile-levels (between 0 and 1).

Sample-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column sample_id of type numeric with sample indices.

For more information see the vignettes and the example data (example_quantile, example_continuous, example_integer, example_point(), and example_binary).

Forecast unit

In order to score forecasts, scoringutils needs to know which of the rows of the data belong together and jointly form a single forecasts. This is easy e.g. for point forecast, where there is one row per forecast. For quantile or sample-based forecasts, however, there are multiple rows that belong to single forecast.

The forecast unit or unit of a single forecast is then described by the combination of columns that uniquely identify a single forecast. For example, we could have forecasts made by different models in various locations at different time points, each for several weeks into the future. The forecast unit could then be described as forecast_unit = c("model", "location", "forecast_date", "forecast_horizon"). scoringutils automatically tries to determine the unit of a single forecast. It uses all existing columns for this, which means that no columns must be present that are unrelated to the forecast unit. As a very simplistic example, if you had an additional row, "even", that is one if the row number is even and zero otherwise, then this would mess up scoring as scoringutils then thinks that this column was relevant in defining the forecast unit.

In order to avoid issues, we recommend using the function set_forecast_unit() to determine the forecast unit manually. The function simply drops unneeded columns, while making sure that all necessary, 'protected columns' like "predicted" or "observed" are retained.

References

Bosse NI, Gruson H, Cori A, van Leeuwen E, Funk S, Abbott S (2022) Evaluating Forecasts with scoringutils in R. doi:10.48550/arXiv.2205.07090

Author

Nikos Bosse nikosbosse@gmail.com

Examples

library(magrittr) # pipe operator
data.table::setDTthreads(1) # only needed to avoid issues on CRAN

validated <- as_forecast(example_quantile)
#> Some rows containing NA values may be removed. This is fine if not unexpected.
score(validated) %>%
  summarise_scores(by = c("model", "target_type"))
#> Some rows containing NA values may be removed. This is fine if not unexpected.
#>                    model target_type         wis overprediction underprediction
#> 1: EuroCOVIDhub-ensemble       Cases 17943.82383   10043.121943     4237.177310
#> 2: EuroCOVIDhub-baseline       Cases 28483.57465   14096.100883    10284.972826
#> 3:  epiforecasts-EpiNow2       Cases 20831.55662   11906.823030     3260.355639
#> 4: EuroCOVIDhub-ensemble      Deaths    41.42249       7.138247        4.103261
#> 5: EuroCOVIDhub-baseline      Deaths   159.40387      65.899117        2.098505
#> 6:       UMass-MechBayes      Deaths    52.65195       8.978601       16.800951
#> 7:  epiforecasts-EpiNow2      Deaths    66.64282      18.892583       15.893314
#>    dispersion        bias coverage_50 coverage_90 coverage_deviation
#> 1: 3663.52458 -0.05640625   0.3906250   0.8046875        -0.10230114
#> 2: 4102.50094  0.09796875   0.3281250   0.8203125        -0.11721591
#> 3: 5664.37795 -0.07890625   0.4687500   0.7890625        -0.06963068
#> 4:   30.18099  0.07265625   0.8750000   1.0000000         0.20380682
#> 5:   91.40625  0.33906250   0.6640625   1.0000000         0.12142045
#> 6:   26.87239 -0.02234375   0.4609375   0.8750000        -0.02488636
#> 7:   31.85692 -0.00512605   0.4201681   0.9075630        -0.04520244
#>      ae_median
#> 1: 24101.07031
#> 2: 38473.60156
#> 3: 27923.81250
#> 4:    53.13281
#> 5:   233.25781
#> 6:    78.47656
#> 7:   104.74790

# set forecast unit manually (to avoid issues with scoringutils trying to
# determine the forecast unit automatically), check forecasts before scoring
example_quantile %>%
  set_forecast_unit(
    c("location", "target_end_date", "target_type", "horizon", "model")
  ) %>%
  as_forecast() %>%
  score()
#> Some rows containing NA values may be removed. This is fine if not unexpected.
#> Some rows containing NA values may be removed. This is fine if not unexpected.
#>      location target_end_date target_type horizon                 model
#>   1:       DE      2021-05-08       Cases       1 EuroCOVIDhub-ensemble
#>   2:       DE      2021-05-08       Cases       1 EuroCOVIDhub-baseline
#>   3:       DE      2021-05-08       Cases       1  epiforecasts-EpiNow2
#>   4:       DE      2021-05-08      Deaths       1 EuroCOVIDhub-ensemble
#>   5:       DE      2021-05-08      Deaths       1 EuroCOVIDhub-baseline
#>  ---                                                                   
#> 883:       IT      2021-07-24      Deaths       2 EuroCOVIDhub-baseline
#> 884:       IT      2021-07-24      Deaths       3       UMass-MechBayes
#> 885:       IT      2021-07-24      Deaths       2       UMass-MechBayes
#> 886:       IT      2021-07-24      Deaths       3  epiforecasts-EpiNow2
#> 887:       IT      2021-07-24      Deaths       2  epiforecasts-EpiNow2
#>               wis overprediction underprediction  dispersion  bias coverage_50
#>   1:  7990.854783   2.549870e+03       0.0000000 5440.985217  0.50        TRUE
#>   2: 16925.046957   1.527583e+04       0.0000000 1649.220870  0.95       FALSE
#>   3: 25395.960870   1.722226e+04       0.0000000 8173.700000  0.90       FALSE
#>   4:    53.880000   0.000000e+00       0.6086957   53.271304 -0.10        TRUE
#>   5:    46.793043   2.130435e+00       0.0000000   44.662609  0.30        TRUE
#>  ---                                                                          
#> 883:    80.336957   3.608696e+00       0.0000000   76.728261  0.20        TRUE
#> 884:     4.881739   4.347826e-02       0.0000000    4.838261  0.10        TRUE
#> 885:    25.581739   1.782609e+01       0.0000000    7.755652  0.80       FALSE
#> 886:    19.762609   5.478261e+00       0.0000000   14.284348  0.50        TRUE
#> 887:    66.161739   4.060870e+01       0.0000000   25.553043  0.90       FALSE
#>      coverage_90 coverage_deviation ae_median
#>   1:        TRUE         0.05181818     12271
#>   2:       FALSE        -0.40272727     25620
#>   3:        TRUE        -0.31181818     44192
#>   4:        TRUE         0.41545455        14
#>   5:        TRUE         0.23363636        15
#>  ---                                         
#> 883:        TRUE         0.32454545        53
#> 884:        TRUE         0.41545455         1
#> 885:        TRUE        -0.22090909        46
#> 886:        TRUE         0.05181818        26
#> 887:        TRUE        -0.31181818       108

# forecast formats with different metrics
if (FALSE) {
score(example_binary)
score(example_quantile)
score(example_point)
score(example_integer)
score(example_continuous)
}