scoringutils

This package is designed to help with assessing the quality of predictions. It provides a collection of proper scoring rules and metrics as well that can be accessed independently or collectively through higher-level wrapper functions.

Predicitions can be either probabilistic forecasts (generally predictive samples generated by Markov Chain Monte Carlo procedures) or point forecasts. The type of the predictions and the true values can be either continuous, integer, or binary.

A collection of different metrics and scoring rules can be accessed through the function eval_forecasts with the correct prediction type (probabilistic/point prediction, and continuous/integer/binary) specified.

The following functions can also be accessed directly:

probabilistic - integer

pit_int

bias

sharpness

crps

dss

probabilistic - binary

brier_score

probabilistic - continuous

pit

bias

sharpness

crps

logs

dss

point prediction - integer

.

point prediction - binary

.

point prediction - continuous

.