scoringutils: Utilities for Scoring and Assessing Predictions

Facilitate the evaluation of forecasts in a convenient framework based on data.table. It allows user to to check their forecasts and diagnose issues, to visualise forecasts and missing data, to transform data before scoring, to handle missing forecasts, to aggregate scores, and to visualise the results of the evaluation. The package mostly focuses on the evaluation of probabilistic forecasts and allows evaluating several different forecast types and input formats. Find more information about the package in the Vignettes as well as in the accompanying paper, doi:10.48550/arXiv.2205.07090 .

See also

Author

Maintainer: Nikos Bosse nikosbosse@gmail.com (ORCID)

Authors:

Sam Abbott contact@samabbott.co.uk (ORCID)

Hugo Gruson hugo.gruson+R@normalesup.org (ORCID)

Sebastian Funk sebastian.funk@lshtm.ac.uk

Other contributors:

Johannes Bracher johannes.bracher@kit.edu (ORCID) [contributor]

Toshiaki Asakura toshiaki.asa9ra@gmail.com (ORCID) [contributor]

James Mba Azam james.azam@lshtm.ac.uk (ORCID) [contributor]

Michael Chirico michaelchirico4@gmail.com (ORCID) [contributor]