Squared Error of the Mean (Sample-based Version)

Squared error of the mean calculated as

$$ \textrm{mean}(\textrm{true\_value} - \textrm{prediction})^2 $$

Usage

se_mean_sample(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n.

Value

vector with the scoring values

See also

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
se_mean_sample(true_values, predicted_values)
#>  [1] 0.646984120 0.568541092 0.626308214 3.390074770 2.259230119 2.654205973
#>  [7] 5.655789843 5.067594025 1.255040987 2.570863646 1.469491988 0.830953748
#> [13] 1.566408518 1.507974309 0.004865038 0.404408308 0.092724548 0.052057889
#> [19] 3.407721726 0.562258033 5.418754993 0.067073894 0.032783937 0.528665931
#> [25] 1.511632216 0.101762998 6.367054867 0.902845543 3.138834918 3.209666529