Set unit of a single forecast manually

Helper function to set the unit of a single forecast (i.e. the combination of columns that uniquely define a single forecast) manually. This simple function keeps the columns specified in forecast_unit (plus additional protected columns, e.g. for observed values, predictions or quantile levels) and removes duplicate rows. If not done manually, scoringutils attempts to determine the unit of a single forecast automatically by simply assuming that all column names are relevant to determine the forecast unit. This may lead to unexpected behaviour, so setting the forecast unit explicitly can help make the code easier to debug and easier to read. When used as part of a workflow, set_forecast_unit() can be directly piped into as_forecast() to check everything is in order.

Usage

set_forecast_unit(data, forecast_unit)

Arguments

data

A data.frame or data.table with predicted and observed values.

forecast_unit

Character vector with the names of the columns that uniquely identify a single forecast.

Value

A data.table with only those columns kept that are relevant to scoring or denote the unit of a single forecast as specified by the user.

Examples

set_forecast_unit(
  example_quantile,
  c("location", "target_end_date", "target_type", "horizon", "model")
)
#>        observed quantile predicted location target_end_date target_type horizon
#>     1:   127300       NA        NA       DE      2021-01-02       Cases      NA
#>     2:     4534       NA        NA       DE      2021-01-02      Deaths      NA
#>     3:   154922       NA        NA       DE      2021-01-09       Cases      NA
#>     4:     6117       NA        NA       DE      2021-01-09      Deaths      NA
#>     5:   110183       NA        NA       DE      2021-01-16       Cases      NA
#>    ---                                                                         
#> 20541:       78    0.850       352       IT      2021-07-24      Deaths       2
#> 20542:       78    0.900       397       IT      2021-07-24      Deaths       2
#> 20543:       78    0.950       499       IT      2021-07-24      Deaths       2
#> 20544:       78    0.975       611       IT      2021-07-24      Deaths       2
#> 20545:       78    0.990       719       IT      2021-07-24      Deaths       2
#>                       model
#>     1:                 <NA>
#>     2:                 <NA>
#>     3:                 <NA>
#>     4:                 <NA>
#>     5:                 <NA>
#>    ---                     
#> 20541: epiforecasts-EpiNow2
#> 20542: epiforecasts-EpiNow2
#> 20543: epiforecasts-EpiNow2
#> 20544: epiforecasts-EpiNow2
#> 20545: epiforecasts-EpiNow2