Validate input data

Methods for the different classes run validate_general(), which performs checks that are the same for all forecast types and then perform specific checks for the specific forecast type.

Usage

validate_forecast(data, forecast_type = NULL, ...)

# S3 method for forecast_quantile
validate_forecast(data, forecast_type = NULL, ...)

# S3 method for forecast_sample
validate_forecast(data, forecast_type = NULL, ...)

Arguments

data

A data.frame (or similar) with predicted and observed values. See as_forecast() for additional information on input formats.

forecast_type

(optional) The forecast type you expect the forecasts to have. If the forecast type as determined by scoringutils based on the input does not match this, an error will be thrown. If NULL (the default), the forecast type will be inferred from the data.

...

additional arguments

Value

Depending on the forecast type, an object of class forecast_binary, forecast_point, forecast_sample or forecast_quantile.

Forecast types and input format

Various different forecast types / forecast formats are supported. At the moment, those are

point forecasts

binary forecasts ("soft binary classification")

Probabilistic forecasts in a quantile-based format (a forecast is represented as a set of predictive quantiles)

Probabilistic forecasts in a sample-based format (a forecast is represented as a set of predictive samples)

Forecast types are determined based on the columns present in the input data.

Point forecasts require a column observed of type numeric and a column predicted of type numeric.

Binary forecasts require a column observed of type factor with exactly two levels and a column predicted of type numeric with probabilities, corresponding to the probability that observed is equal to the second factor level. See details here for more information.

Quantile-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column quantile_level of type numeric with quantile-levels (between 0 and 1).

Sample-based forecasts require a column observed of type numeric, a column predicted of type numeric, and a column sample_id of type numeric with sample indices.

For more information see the vignettes and the example data (example_quantile, example_continuous, example_integer, example_point(), and example_binary).

Examples

forecast <- as_forecast(example_binary)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
validate_forecast(forecast)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#> Forecast type:
#> binary
#> Forecast unit:
#> location, location_name, target_end_date, target_type, forecast_date, model,
#> and horizon
#> 
#>       location location_name target_end_date target_type forecast_date
#>         <char>        <char>          <Date>      <char>        <Date>
#>    1:       DE       Germany      2021-01-02       Cases          <NA>
#>    2:       DE       Germany      2021-01-02      Deaths          <NA>
#>    3:       DE       Germany      2021-01-09       Cases          <NA>
#>    4:       DE       Germany      2021-01-09      Deaths          <NA>
#>    5:       DE       Germany      2021-01-16       Cases          <NA>
#>   ---                                                                 
#> 1027:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 1028:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 1029:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 1030:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 1031:       IT         Italy      2021-07-24      Deaths    2021-07-12
#>                       model horizon predicted observed
#>                      <char>   <num>     <num>   <fctr>
#>    1:                  <NA>      NA        NA     <NA>
#>    2:                  <NA>      NA        NA     <NA>
#>    3:                  <NA>      NA        NA     <NA>
#>    4:                  <NA>      NA        NA     <NA>
#>    5:                  <NA>      NA        NA     <NA>
#>   ---                                                 
#> 1027: EuroCOVIDhub-baseline       2     0.250        0
#> 1028:       UMass-MechBayes       3     0.475        0
#> 1029:       UMass-MechBayes       2     0.450        0
#> 1030:  epiforecasts-EpiNow2       3     0.375        0
#> 1031:  epiforecasts-EpiNow2       2     0.300        0