Absolute Error of the Median (Quantile-based Version)

Absolute error of the median calculated as

$$ abs(true_value - median_prediction) $$

ae_median_quantile(true_values, predictions, quantiles = NULL, verbose = TRUE)

Arguments

true_values

A vector with the true observed values of size n

predictions

numeric vector with predictions, corresponding to the quantiles in a second vector, quantiles.

quantiles

numeric vector that denotes the quantile for the values in predictions. Only those predictions where quantiles == 0.5 will be kept. If quantiles is NULL, then all predictions and true_values will be used (this is then the same as abs_error())

verbose

logical, return a warning is something unexpected happens

Value

vector with the scoring values

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
ae_median_quantile(true_values, predicted_values, quantiles = 0.5)
#>  [1] 0.31090232 0.33777632 1.10516257 1.71937515 2.07422917 0.75583716
#>  [7] 1.68871511 0.93417175 0.64328729 0.34465247 0.43137500 1.63132219
#> [13] 0.48144362 2.99663527 2.81456153 0.21128078 1.65048824 0.47474630
#> [19] 0.21399991 0.94402894 1.22232188 0.80723497 0.04445436 0.13225600
#> [25] 1.15142948 1.16394940 0.26229646 0.89752993 1.23965683 0.47933710