Absolute Error of the Median (Sample-based Version)

Absolute error of the median calculated as

$$ abs(true_value - median_prediction) $$

ae_median_sample(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n

Value

vector with the scoring values

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
ae_median_sample(true_values, predicted_values)
#>  [1] 3.7628868 0.8185383 0.6144328 2.2613157 1.6806267 0.2809775 0.1840405
#>  [8] 0.1792891 1.2454297 4.0252060 1.5804432 0.2173886 1.9065063 0.4213602
#> [15] 0.1069744 2.1675463 1.8126443 0.8594406 2.9625172 1.2683812 0.1957751
#> [22] 1.4874834 1.8041212 0.5433466 0.5978937 0.0109446 1.6198748 1.2679847
#> [29] 3.5595198 0.8318742