Absolute Error of the Median (Sample-based Version)

Absolute error of the median calculated as

$$ abs(true_value - median_prediction) $$

ae_median_sample(true_values, predictions)

Arguments

true_values
A vector with the true observed values of size n

predictions
nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n

Value

vector with the scoring values

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
ae_median_sample(true_values, predicted_values)
#>  [1] 1.25728308 2.37277306 0.23615771 1.67012595 0.98211657 0.07111145
#>  [7] 0.19835287 1.63979528 2.19329734 0.17106903 1.91213754 1.95402344
#> [13] 1.29109628 1.12144968 1.21642659 0.84450490 0.91631200 1.03121139
#> [19] 0.31887378 0.73472164 0.55083953 1.36803337 0.25217732 0.26628408
#> [25] 0.03761679 2.06708539 1.30003812 1.79296333 0.28488936 0.49732308