Determines bias of forecasts

Determines bias from predictive Monte-Carlo samples. The function automatically recognises, whether forecasts are continuous or integer valued and adapts the Bias function accordingly.

bias(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector of length n with the biases of the predictive samples with respect to the true values.

Details

For continuous forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - 2 * (P_t (x_t)) $$

where \(P_t\) is the empirical cumulative distribution function of the prediction for the true value \(x_t\). Computationally, \(P_t (x_t)\) is just calculated as the fraction of predictive samples for \(x_t\) that are smaller than \(x_t\).

For integer valued forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - (P_t (x_t) + P_t (x_t + 1)) $$

to adjust for the integer nature of the forecasts.

In both cases, Bias can assume values between -1 and 1 and is 0 ideally.

References

The integer valued Bias function is discussed in Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15 Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, et al. (2019) Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15. PLOS Computational Biology 15(2): e1006785. doi:10.1371/journal.pcbi.1006785

Author

Nikos Bosse nikosbosse@gmail.com

Examples

## integer valued forecasts
true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
bias(true_values, predictions)
#>  [1] -0.925 -0.030 -0.130 -0.090 -0.745 -0.040 -0.070 -0.035  0.265  0.900
#> [11] -0.320  0.370  0.390  0.385 -0.350 -0.695 -0.570  0.420  0.005  0.535
#> [21]  0.220  0.055 -0.180 -0.995  0.060 -0.830 -0.895 -0.090 -0.670  0.350

## continuous forecasts
true_values <- rnorm(30, mean = 1:30)
predictions <- replicate(200, rnorm(30, mean = 1:30))
bias(true_values, predictions)
#>  [1] -0.64  0.87 -0.79  0.34  0.33  0.90  0.27 -0.08 -0.94  0.02 -0.27 -0.96
#> [13] -0.87  0.81 -0.67  0.94  0.79  0.13  0.13  0.52  0.18 -0.26  0.98  0.08
#> [25]  0.51  0.67 -0.56  0.21 -0.79  0.81