Determines Bias Of Forecasts

Determines bias from predictive Monte-Carlo samples. The function automatically recognises, whether forecasts are continuous or integer valued and adapts the Bias function accordingly.

Usage

bias_sample(observed, predicted)

Arguments

observed

A vector with observed values of size n

predicted

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predicted can just be a vector of size n.

Value

vector of length n with the biases of the predictive samples with respect to the observed values.

Details

For continuous forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - 2 * (P_t (x_t)) $$

where \(P_t\) is the empirical cumulative distribution function of the prediction for the observed value \(x_t\). Computationally, \(P_t (x_t)\) is just calculated as the fraction of predictive samples for \(x_t\) that are smaller than \(x_t\).

For integer valued forecasts, Bias is measured as

$$ B_t (P_t, x_t) = 1 - (P_t (x_t) + P_t (x_t + 1)) $$

to adjust for the integer nature of the forecasts.

In both cases, Bias can assume values between -1 and 1 and is 0 ideally.

References

The integer valued Bias function is discussed in Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15 Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, et al. (2019) Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15. PLOS Computational Biology 15(2): e1006785. doi:10.1371/journal.pcbi.1006785

Author

Nikos Bosse nikosbosse@gmail.com

Examples

## integer valued forecasts
observed <- rpois(30, lambda = 1:30)
predicted <- replicate(200, rpois(n = 30, lambda = 1:30))
bias_sample(observed, predicted)
#>  [1] -0.890  0.840 -0.480  0.245 -0.770 -0.405 -0.365 -0.415  0.015  0.610
#> [11] -0.500  0.135  0.440  0.115  0.840 -0.455  0.015  0.340 -0.450 -0.860
#> [21] -0.895 -0.195  0.925  0.220  0.655  0.750 -0.645 -0.760  0.860  0.270

## continuous forecasts
observed <- rnorm(30, mean = 1:30)
predicted <- replicate(200, rnorm(30, mean = 1:30))
bias_sample(observed, predicted)
#>  [1]  0.06 -0.43 -0.15  0.95 -0.44  0.43 -0.79  0.03 -0.90  0.26 -0.32 -0.68
#> [13] -0.04  0.25 -0.15  0.35  0.42  0.70 -0.03 -0.99  0.31 -0.34  0.63  0.16
#> [25]  0.35  0.90  0.47 -0.76 -0.85  0.78