Correlation Between Metrics

Calculate the correlation between different metrics for a data.frame of scores as produced by score().

Usage

correlation(scores, metrics = NULL, digits = NULL)

Arguments

scores

A data.table of scores as produced by score().

metrics

A character vector with the metrics to show. If set to NULL (default), all metrics present in scores will be shown

digits

A number indicating how many decimal places the result should be rounded to. By default (digits = NULL) no rounding takes place.

Value

A data.table with correlations for the different metrics

Examples

scores <- score(example_quantile)
correlation(scores, digits = 2)
#>      wis overprediction underprediction dispersion  bias coverage_deviation
#> 1:  1.00           0.94            0.28       0.46  0.11              -0.34
#> 2:  0.94           1.00           -0.03       0.32  0.22              -0.25
#> 3:  0.28          -0.03            1.00       0.15 -0.35              -0.33
#> 4:  0.46           0.32            0.15       1.00  0.11              -0.12
#> 5:  0.11           0.22           -0.35       0.11  1.00               0.06
#> 6: -0.34          -0.25           -0.33      -0.12  0.06               1.00
#> 7:  0.99           0.90            0.34       0.54  0.10              -0.38
#>    ae_median             metric
#> 1:      0.99                wis
#> 2:      0.90     overprediction
#> 3:      0.34    underprediction
#> 4:      0.54         dispersion
#> 5:      0.10               bias
#> 6:     -0.38 coverage_deviation
#> 7:      1.00          ae_median