Customises a metric function with additional arguments.

This function takes a metric function and additional arguments, and returns a new function that includes the additional arguments when calling the original metric function.

This is the expected way to pass additional arguments to a metric when evaluating a forecast using score(): To evaluate a forecast using a metric with an additional argument, you need to create a custom version of the scoring function with the argument included. You then need to create an updated version of the list of scoring functions that includes your customised metric and pass this to score().

Usage

customise_metric(metric, ...)

customize_metric(metric, ...)

Arguments

metric

The metric function to be customised.

...

Additional arguments to be included when calling the metric function.

Value

A customised metric function.

Examples

# Create a customised metric function
custom_metric <- customise_metric(mean, na.rm = TRUE)

# Use the customised metric function
values <- c(1, 2, NA, 4, 5)
custom_metric(values)
#> [1] 3

# Updating metrics list to calculate 70% coverage in `score()`
interval_coverage_70 <- customise_metric(
  interval_coverage, interval_range = 70
)
updated_metrics <- c(
  metrics_quantile(),
  "interval_coverage_70" = interval_coverage_70
)
score(
  as_forecast(example_quantile),
  metrics = updated_metrics
)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
#>      location target_end_date target_type location_name forecast_date
#>        <char>          <Date>      <char>        <char>        <Date>
#>   1:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   2:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   3:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   4:       DE      2021-05-08      Deaths       Germany    2021-05-03
#>   5:       DE      2021-05-08      Deaths       Germany    2021-05-03
#>  ---                                                                 
#> 883:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 884:       IT      2021-07-24      Deaths         Italy    2021-07-05
#> 885:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 886:       IT      2021-07-24      Deaths         Italy    2021-07-05
#> 887:       IT      2021-07-24      Deaths         Italy    2021-07-12
#>                      model horizon          wis overprediction underprediction
#>                     <char>   <num>        <num>          <num>           <num>
#>   1: EuroCOVIDhub-ensemble       1  7990.854783   2.549870e+03       0.0000000
#>   2: EuroCOVIDhub-baseline       1 16925.046957   1.527583e+04       0.0000000
#>   3:  epiforecasts-EpiNow2       1 25395.960870   1.722226e+04       0.0000000
#>   4: EuroCOVIDhub-ensemble       1    53.880000   0.000000e+00       0.6086957
#>   5: EuroCOVIDhub-baseline       1    46.793043   2.130435e+00       0.0000000
#>  ---                                                                          
#> 883: EuroCOVIDhub-baseline       2    80.336957   3.608696e+00       0.0000000
#> 884:       UMass-MechBayes       3     4.881739   4.347826e-02       0.0000000
#> 885:       UMass-MechBayes       2    25.581739   1.782609e+01       0.0000000
#> 886:  epiforecasts-EpiNow2       3    19.762609   5.478261e+00       0.0000000
#> 887:  epiforecasts-EpiNow2       2    66.161739   4.060870e+01       0.0000000
#>       dispersion  bias interval_coverage_50 interval_coverage_90
#>            <num> <num>               <lgcl>               <lgcl>
#>   1: 5440.985217  0.50                 TRUE                 TRUE
#>   2: 1649.220870  0.95                FALSE                FALSE
#>   3: 8173.700000  0.90                FALSE                 TRUE
#>   4:   53.271304 -0.10                 TRUE                 TRUE
#>   5:   44.662609  0.30                 TRUE                 TRUE
#>  ---                                                            
#> 883:   76.728261  0.20                 TRUE                 TRUE
#> 884:    4.838261  0.10                 TRUE                 TRUE
#> 885:    7.755652  0.80                FALSE                 TRUE
#> 886:   14.284348  0.50                 TRUE                 TRUE
#> 887:   25.553043  0.90                FALSE                 TRUE
#>      interval_coverage_deviation ae_median interval_coverage_70
#>                            <num>     <num>               <lgcl>
#>   1:                  0.05181818     12271                 TRUE
#>   2:                 -0.40272727     25620                FALSE
#>   3:                 -0.31181818     44192                FALSE
#>   4:                  0.41545455        14                 TRUE
#>   5:                  0.23363636        15                 TRUE
#>  ---                                                           
#> 883:                  0.32454545        53                 TRUE
#> 884:                  0.41545455         1                 TRUE
#> 885:                 -0.22090909        46                FALSE
#> 886:                  0.05181818        26                 TRUE
#> 887:                 -0.31181818       108                FALSE