Dawid-Sebastiani Score

Wrapper around the dss_sample function from the scoringRules package.

dss(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts withscoringRules, https://arxiv.org/pdf/1709.04743.pdf

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
dss(true_values, predictions)
#>  [1] -0.0253176  1.2516285  1.7061818  1.5931555  2.5008231  3.3368390
#>  [7]  2.1839105  2.1540394  2.2353898  2.4225905  2.4379901  2.4220423
#> [13]  3.3458300  2.4553086  3.0340440  4.3615757  2.9847060  3.7196156
#> [19]  4.2315519  4.2435457  3.1587019  4.8912272  4.7102695  5.4683447
#> [25]  3.5120013  3.2156256  3.3074361  3.5040447  3.5375029  7.7820358