Dawid-Sebastiani Score

Wrapper around the dss_sample function from the scoringRules package.

dss(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts withscoringRules, https://arxiv.org/pdf/1709.04743.pdf

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
dss(true_values, predictions)
#>  [1] 1.231836 1.137328 1.152200 2.329691 1.764735 1.999164 2.418588 2.361423
#>  [9] 2.334183 2.486805 3.922205 2.708305 3.117076 3.724975 2.932596 3.199935
#> [17] 2.915362 4.406633 2.872968 3.224666 3.038291 3.062514 5.104938 5.890354
#> [25] 3.625141 5.886210 3.752472 7.751237 3.462935 4.017399