Calculate correlation between metrics

Calculate the correlation between different metrics for a data.frame of scores as produced by score().

Usage

get_correlations(scores, metrics = get_metrics(scores), digits = NULL, ...)

Arguments

scores

An object of class scores (a data.table with scores and an additional attribute metrics as produced by score()).

metrics

A character vector with the metrics to show. If set to NULL (default), all metrics present in scores will be shown.

digits

A number indicating how many decimal places the result should be rounded to. By default (digits = NULL) no rounding takes place.

...

Additional arguments to pass down to cor().

Value

An object of class scores (a data.table with an additional attribute metrics holding the names of the scores) with correlations between different metrics

Examples

scores <- score(as_forecast(example_quantile))
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.
get_correlations(scores, digits = 2)
#>      wis overprediction underprediction dispersion  bias interval_coverage_50
#>    <num>          <num>           <num>      <num> <num>                <num>
#> 1:  1.00           0.94            0.28       0.46  0.11                -0.21
#> 2:  0.94           1.00           -0.03       0.32  0.22                -0.15
#> 3:  0.28          -0.03            1.00       0.15 -0.35                -0.21
#> 4:  0.46           0.32            0.15       1.00  0.11                -0.09
#> 5:  0.11           0.22           -0.35       0.11  1.00                 0.01
#> 6: -0.21          -0.15           -0.21      -0.09  0.01                 1.00
#> 7: -0.41          -0.32           -0.36      -0.09  0.10                 0.37
#> 8: -0.34          -0.25           -0.33      -0.12  0.06                 0.85
#> 9:  0.99           0.90            0.34       0.54  0.10                -0.25
#>    interval_coverage_90 interval_coverage_deviation ae_median
#>                   <num>                       <num>     <num>
#> 1:                -0.41                       -0.34      0.99
#> 2:                -0.32                       -0.25      0.90
#> 3:                -0.36                       -0.33      0.34
#> 4:                -0.09                       -0.12      0.54
#> 5:                 0.10                        0.06      0.10
#> 6:                 0.37                        0.85     -0.25
#> 7:                 1.00                        0.64     -0.41
#> 8:                 0.64                        1.00     -0.38
#> 9:                -0.41                       -0.38      1.00
#>                         metric
#>                         <char>
#> 1:                         wis
#> 2:              overprediction
#> 3:             underprediction
#> 4:                  dispersion
#> 5:                        bias
#> 6:        interval_coverage_50
#> 7:        interval_coverage_90
#> 8: interval_coverage_deviation
#> 9:                   ae_median