Interval Score

Proper Scoring Rule to score quantile predictions, following Gneiting and Raftery (2007). Smaller values are better.

The score is computed as

$$ \textrm{score} = (\textrm{upper} - \textrm{lower}) + \frac{2}{\alpha}(\textrm{lower} - \textrm{true\_value}) * \mathbf{1}(\textrm{true\_value} < \textrm{lower}) + \frac{2}{\alpha}(\textrm{true\_value} - \textrm{upper}) * \mathbf{1}(\textrm{true\_value} > \textrm{upper}) $$ where \(\mathbf{1}()\) is the indicator function and indicates how much is outside the prediction interval. \(\alpha\) is the decimal value that indicates how much is outside the prediction interval.

To improve usability, the user is asked to provide an interval range in percentage terms, i.e. interval_range = 90 (percent) for a 90 percent prediction interval. Correspondingly, the user would have to provide the 5% and 95% quantiles (the corresponding alpha would then be 0.1). No specific distribution is assumed, but the range has to be symmetric (i.e you can't use the 0.1 quantile as the lower bound and the 0.7 quantile as the upper). Non-symmetric quantiles can be scored using the function quantile_score().

Usage

interval_score(
  true_values,
  lower,
  upper,
  interval_range,
  weigh = TRUE,
  separate_results = FALSE
)

Arguments

true_values

A vector with the true observed values of size n

lower

vector of size n with the prediction for the lower quantile of the given range

upper

vector of size n with the prediction for the upper quantile of the given range

interval_range

the range of the prediction intervals. i.e. if you're forecasting the 0.05 and 0.95 quantile, the interval_range would be 90. Can be either a single number or a vector of size n, if the range changes for different forecasts to be scored. This corresponds to (100-alpha)/100 in Gneiting and Raftery (2007). Internally, the range will be transformed to alpha.

weigh

if TRUE, weigh the score by alpha / 2, so it can be averaged into an interval score that, in the limit, corresponds to CRPS. Alpha is the decimal value that represents how much is outside a central prediction interval (e.g. for a 90 percent central prediction interval, alpha is 0.1) Default: TRUE.

separate_results

if TRUE (default is FALSE), then the separate parts of the interval score (dispersion penalty, penalties for over- and under-prediction get returned as separate elements of a list). If you want a data.frame instead, simply call as.data.frame() on the output.

Value

vector with the scoring values, or a list with separate entries if separate_results is TRUE.

References

Strictly Proper Scoring Rules, Prediction,and Estimation, Tilmann Gneiting and Adrian E. Raftery, 2007, Journal of the American Statistical Association, Volume 102, 2007 - Issue 477

Evaluating epidemic forecasts in an interval format, Johannes Bracher, Evan L. Ray, Tilmann Gneiting and Nicholas G. Reich, https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1008618 # nolint

Examples

true_values <- rnorm(30, mean = 1:30)
interval_range <- rep(90, 30)
alpha <- (100 - interval_range) / 100
lower <- qnorm(alpha / 2, rnorm(30, mean = 1:30))
upper <- qnorm((1 - alpha / 2), rnorm(30, mean = 1:30))

interval_score(
  true_values = true_values,
  lower = lower,
  upper = upper,
  interval_range = interval_range
)
#>  [1] 0.21658067 1.28381486 1.65065309 0.58076291 0.12111804 0.14706138
#>  [7] 0.21467759 0.25479793 1.37265205 0.19623593 0.15114765 0.18187711
#> [13] 0.09334673 0.03537728 0.16882400 0.29176782 0.78471082 0.24762664
#> [19] 0.27654729 0.28412533 0.19843789 1.12189772 0.23337973 0.06172599
#> [25] 1.40020869 0.16821258 0.23291196 0.11106442 0.14183550 0.02742978

# example with missing values and separate results
interval_score(
  true_values = c(true_values, NA),
  lower = c(lower, NA),
  upper = c(NA, upper),
  separate_results = TRUE,
  interval_range = 90
)
#> $interval_score
#>  [1]         NA 0.07039680 3.84297804 2.30165658 1.17431775 0.04029965
#>  [7] 0.20269017 0.13188946 0.19248313 1.13717744 0.14886306 0.18302871
#> [13] 0.05256620 0.54998129 1.43812269 0.15865956 0.14561278 1.28981993
#> [19] 0.21401701 0.19746655 0.17993965 1.21669936 3.01814882 0.09092366
#> [25] 1.76438536 1.73954105 0.12523843 0.13929293 2.10256783 0.11898420
#> [31]         NA
#> 
#> $dispersion
#>  [1]           NA  0.070396796  0.026187679  0.115401200  0.062725449
#>  [6]  0.040299647  0.202690174  0.131889456  0.192483127  0.032527887
#> [11]  0.148863061  0.183028711  0.051222210  0.001264727  0.052474888
#> [16]  0.158659565  0.145612784  0.063478537  0.214017012  0.197466555
#> [21]  0.179939647  0.060905968  0.072540111  0.090923663  0.079489729
#> [26]  0.083959852  0.125238426  0.139292927 -0.015443536  0.118984200
#> [31]           NA
#> 
#> $underprediction
#>  [1]          NA 0.000000000 3.816790364 2.186255380 1.111592296 0.000000000
#>  [7] 0.000000000 0.000000000 0.000000000 1.104649550 0.000000000 0.000000000
#> [13] 0.001343995 0.548716558 1.385647806 0.000000000 0.000000000 1.226341397
#> [19] 0.000000000 0.000000000 0.000000000 0.000000000 2.945608705 0.000000000
#> [25] 1.684895632 1.655581197 0.000000000 0.000000000 2.118011363 0.000000000
#> [31]          NA
#> 
#> $overprediction
#>  [1] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [17] 0.000000 0.000000 0.000000 0.000000 0.000000 1.155793 0.000000 0.000000
#> [25] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000       NA
#>