Interval Score

Proper Scoring Rule to score quantile predictions, following Gneiting and Raftery (2007). Smaller values are better.

The score is computed as

$$ score = (upper - lower) + 2/alpha * (lower - true_value) * 1(true_values < lower) + 2/alpha * (true_value - upper) * 1(true_value > upper) $$ where $1()$ is the indicator function and alpha is the decimal value that indicates how much is outside the prediction interval. To improve usability, the user is asked to provide an interval range in percentage terms, i.e. interval_range = 90 (percent) for a 90 percent prediction interval. Correspondingly, the user would have to provide the 5% and 95% quantiles (the corresponding alpha would then be 0.1). No specific distribution is assumed, but the range has to be symmetric (i.e you can't use the 0.1 quantile as the lower bound and the 0.7 quantile as the upper).

The interval score is a proper scoring rule that scores a quantile forecast

interval_score(
  true_values,
  lower,
  upper,
  interval_range,
  weigh = TRUE,
  separate_results = FALSE
)

Arguments

true_values

A vector with the true observed values of size n

lower

vector of size n with the lower quantile of the given range

upper

vector of size n with the upper quantile of the given range

interval_range

the range of the prediction intervals. i.e. if you're forecasting the 0.05 and 0.95 quantile, the interval_range would be 90. Can be either a single number or a vector of size n, if the range changes for different forecasts to be scored. This corresponds to (100-alpha)/100 in Gneiting and Raftery (2007). Internally, the range will be transformed to alpha.

weigh

if TRUE, weigh the score by alpha / 4, so it can be averaged into an interval score that, in the limit, corresponds to CRPS. Default: FALSE.

separate_results

if TRUE (default is FALSE), then the separate parts of the interval score (sharpness, penalties for over- and under-prediction get returned as separate elements of a list). If you want a `data.frame` instead, simply call `as.data.frmae()` on the output.

Value

vector with the scoring values, or a list with separate entries if separate_results is TRUE.

References

Strictly Proper Scoring Rules, Prediction,and Estimation, Tilmann Gneiting and Adrian E. Raftery, 2007, Journal of the American Statistical Association, Volume 102, 2007 - Issue 477

Evaluating epidemic forecasts in an interval format, Johannes Bracher, Evan L. Ray, Tilmann Gneiting and Nicholas G. Reich, <arXiv:2005.12881v1>

Bracher J, Ray E, Gneiting T, Reich, N (2020) Evaluating epidemic forecasts in an interval format. https://arxiv.org/abs/2005.12881

Examples

true_values <- rnorm(30, mean = 1:30)
interval_range = rep(90, 30)
alpha = (100 - interval_range) / 100
lower = qnorm(alpha/2, rnorm(30, mean = 1:30))
upper = qnorm((1- alpha/2), rnorm(30, mean = 1:30))

interval_score(true_values = true_values,
               lower = lower,
               upper = upper,
               interval_range = interval_range)
#>  [1] 0.17475510 0.26218781 0.17598102 0.33006901 1.61440929 1.09106643
#>  [7] 1.06262988 0.14169507 0.12331364 0.21809653 1.13213246 0.28856420
#> [13] 0.18843522 0.23666969 0.24314851 0.37432371 0.41496659 0.09331451
#> [19] 0.77792993 0.18360269 0.09629221 0.14308403 0.97483883 0.40139823
#> [25] 0.04809485 0.10270602 0.58989888 0.14751415 0.21903844 0.17882267

# example with missing values and separate results
interval_score(true_values = c(true_values, NA),
               lower = c(lower, NA),
               upper = c(NA, upper),
               separate_results = TRUE,
               interval_range = 90)
#> $interval_score
#>  [1]         NA 0.13192476 0.20474578 0.10866468 5.26195041 1.10223565
#>  [7] 1.69833819 0.40092847 0.11317360 0.17972852 0.96580952 0.28198843
#> [13] 0.18932461 0.11552091 0.18117802 0.33631721 0.22862434 0.98969252
#> [19] 0.28017782 0.95329689 0.08580546 0.17988065 0.94488201 0.81530142
#> [25] 0.04407175 1.55025751 0.17366500 0.70421240 0.58058782 0.24740492
#> [31]         NA
#> 
#> $sharpness
#>  [1]           NA  0.131924762  0.204745782  0.108664678 -0.001658324
#>  [6]  0.111890899  0.166541005  0.077869209  0.050758623  0.179728522
#> [11]  0.023374866  0.281988428  0.189324610  0.115520908  0.181178015
#> [16]  0.336317210  0.228624338 -0.012090950  0.127314287  0.049094924
#> [21]  0.085805464  0.052886798  0.030150308  0.062693489  0.044071746
#> [26] -0.036566138  0.173665000  0.055480447  0.193516176  0.135976869
#> [31]           NA
#> 
#> $underprediction
#>  [1]         NA 0.00000000 0.00000000 0.00000000 5.26360874 0.00000000
#>  [7] 1.53179718 0.32305926 0.06241497 0.00000000 0.00000000 0.00000000
#> [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 1.00178347
#> [19] 0.15286353 0.90420196 0.00000000 0.12699385 0.00000000 0.75260794
#> [25] 0.00000000 1.58682365 0.00000000 0.64873195 0.38707165 0.11142805
#> [31]         NA
#> 
#> $overprediction
#>  [1] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9903448 0.0000000
#>  [8] 0.0000000 0.0000000 0.0000000 0.9424347 0.0000000 0.0000000 0.0000000
#> [15] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [22] 0.0000000 0.9147317 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [29] 0.0000000 0.0000000        NA
#>