LogS

Wrapper around the logs_sample function from the scoringRules package. Used to score continuous predictions. While the Log Score is in theory also applicable to integer forecasts, the problem lies in the implementation: The Log Score needs a kernel density estimation, which is not well defined with integer-valued Monte Carlo Samples. The Log Score can be used for specific integer valued probabiliy distributions. See the scoringRules package for more details.

logs(true_values, predictions)

Arguments

true_values
A vector with the true observed values of size n

predictions
nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with the scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts withscoringRules, https://arxiv.org/pdf/1709.04743.pdf

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
logs(true_values, predictions)

#>  [1] 0.5651598 1.3011258 1.5499496 1.9644559 2.0967969 1.8718419 2.0442318
#>  [8] 2.1130667 2.5225856 2.6577304 2.2045652 2.1993289 4.5750503 2.2885058
#> [15] 3.2055484 2.5250942 3.6164554 2.5661784 2.4654972 2.6036100 3.0155768
#> [22] 2.5362107 2.6483001 2.7913424 2.7373976 2.9422397 2.5261268 5.2147598
#> [29] 3.5558491 2.6547767