LogS

Wrapper around the logs_sample() function from the scoringRules package. Used to score continuous predictions. While the Log Score is in theory also applicable to integer forecasts, the problem lies in the implementation: The Log Score needs a kernel density estimation, which is not well defined with integer-valued Monte Carlo Samples. The Log Score can be used for specific integer valued probability distributions. See the scoringRules package for more details.

logs(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples

Value

vector with the scoring values

References

Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic Forecasts with scoringRules, https://arxiv.org/pdf/1709.04743.pdf

Examples

true_values <- rpois(30, lambda = 1:30)
predictions <- replicate(200, rpois(n = 30, lambda = 1:30))
logs(true_values, predictions)
#>  [1] 2.521314 2.099794 1.426211 2.773900 1.781897 3.778564 2.171637 2.099033
#>  [9] 2.201545 4.118327 3.167744 2.515577 2.719607 2.238035 2.682366 2.683720
#> [17] 2.572604 2.875912 2.670528 2.437675 2.607143 2.418438 2.950042 3.053461
#> [25] 2.679063 2.645951 2.692891 2.606936 3.383689 4.671227