Evaluate forecasts

The function score allows automatic scoring of forecasts and wraps the lower level functions in the scoringutils package.

It can be used to score forecasts in a quantile-based, sample-based, or binary format. To obtain an overview of what input is expected, have a look at the example_quantile, example_continuous, example_integer, and example_binary data sets.

You can (and should) check your input using the function check_forecasts() before scoring.

To obtain a quick overview of the evaluation metrics used, have a look at the metrics data included in the package. The column metrics$Name gives an overview of all available metric names that can be computed.

Usage

score(data, metrics = NULL, ...)

Arguments

data

A data.frame or data.table with the predictions and observations. The following columns need to be present:

true_value - the true observed values

prediction - predictions or predictive samples for one true value. (You only don't need to provide a prediction column if you want to score quantile forecasts in a wide range format.)

For integer and continuous forecasts a sample column is needed:

sample - an index to identify the predictive samples in the prediction column generated by one model for one true value. Only necessary for continuous and integer forecasts, not for binary predictions.

For a quantile-format forecast you should provide a column called quantile:

quantile: quantile to which the prediction corresponds

metrics

the metrics you want to have in the output. If NULL (the default), all available metrics will be computed. For a list of available metrics see available_metrics(), or check the metrics data set.

...

additional parameters passed down to lower-level functions. For example, the following arguments can change how weighted interval scores are computed:

count_median_twice that controls how the interval scores for different intervals are summed up. This should be a logical (default is FALSE) that indicates whether or not to count the median twice when summarising. This would conceptually treat the median as a 0% prediction interval, where the median is the lower as well as the upper bound. The alternative is to treat the median as a single quantile forecast instead of an interval. The interval score would then be better understood as an average of quantile scores.)

Value

A data.table with unsummarised scores. There will be one score per quantile or sample, which is usually not desired, so you should always run summarise_scores() on the unsummarised scores.

References

Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, Edmunds WJ (2019) Assessing the performance of real-time epidemic forecasts: A case study of Ebola in the Western Area region of Sierra Leone, 2014-15. PLoS Comput Biol 15(2): e1006785. doi:10.1371/journal.pcbi.1006785

Author

Nikos Bosse nikosbosse@gmail.com

Examples

library(magrittr) # pipe operator

check_forecasts(example_quantile)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#> Your forecasts seem to be for a target of the following type:
#> $target_type
#> [1] "integer"
#> 
#> and in the following format:
#> $prediction_type
#> [1] "quantile"
#> 
#> The unit of a single forecast is defined by:
#> $forecast_unit
#> [1] "location"        "target_end_date" "target_type"     "location_name"  
#> [5] "forecast_date"   "model"           "horizon"        
#> 
#> Cleaned data, rows with NA values in prediction or true_value removed:
#> $cleaned_data
#>        location target_end_date target_type true_value location_name
#>     1:       DE      2021-05-08       Cases     106987       Germany
#>     2:       DE      2021-05-08       Cases     106987       Germany
#>     3:       DE      2021-05-08       Cases     106987       Germany
#>     4:       DE      2021-05-08       Cases     106987       Germany
#>     5:       DE      2021-05-08       Cases     106987       Germany
#>    ---                                                              
#> 20397:       IT      2021-07-24      Deaths         78         Italy
#> 20398:       IT      2021-07-24      Deaths         78         Italy
#> 20399:       IT      2021-07-24      Deaths         78         Italy
#> 20400:       IT      2021-07-24      Deaths         78         Italy
#> 20401:       IT      2021-07-24      Deaths         78         Italy
#>        forecast_date quantile prediction                 model horizon
#>     1:    2021-05-03    0.010      82466 EuroCOVIDhub-ensemble       1
#>     2:    2021-05-03    0.025      86669 EuroCOVIDhub-ensemble       1
#>     3:    2021-05-03    0.050      90285 EuroCOVIDhub-ensemble       1
#>     4:    2021-05-03    0.100      95341 EuroCOVIDhub-ensemble       1
#>     5:    2021-05-03    0.150      99171 EuroCOVIDhub-ensemble       1
#>    ---                                                                
#> 20397:    2021-07-12    0.850        352  epiforecasts-EpiNow2       2
#> 20398:    2021-07-12    0.900        397  epiforecasts-EpiNow2       2
#> 20399:    2021-07-12    0.950        499  epiforecasts-EpiNow2       2
#> 20400:    2021-07-12    0.975        611  epiforecasts-EpiNow2       2
#> 20401:    2021-07-12    0.990        719  epiforecasts-EpiNow2       2
#> 
#> Number of unique values per column per model:
#> $unique_values
#>                    model location target_end_date target_type true_value
#> 1: EuroCOVIDhub-ensemble        4              12           2         96
#> 2: EuroCOVIDhub-baseline        4              12           2         96
#> 3:  epiforecasts-EpiNow2        4              12           2         95
#> 4:       UMass-MechBayes        4              12           1         48
#>    location_name forecast_date quantile prediction horizon
#> 1:             4            11       23       3969       3
#> 2:             4            11       23       3733       3
#> 3:             4            11       23       3903       3
#> 4:             4            11       23       1058       3
#> 
#> $messages
#> [1] "Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected."
#> 
score(example_quantile) %>%
  add_coverage(by = c("model", "target_type")) %>%
  summarise_scores(by = c("model", "target_type"))
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>                    model target_type interval_score dispersion underprediction
#> 1: EuroCOVIDhub-baseline       Cases    28483.57465 4102.50094    10284.972826
#> 2: EuroCOVIDhub-baseline      Deaths      159.40387   91.40625        2.098505
#> 3: EuroCOVIDhub-ensemble       Cases    17943.82383 3663.52458     4237.177310
#> 4: EuroCOVIDhub-ensemble      Deaths       41.42249   30.18099        4.103261
#> 5:       UMass-MechBayes      Deaths       52.65195   26.87239       16.800951
#> 6:  epiforecasts-EpiNow2       Cases    20831.55662 5664.37795     3260.355639
#> 7:  epiforecasts-EpiNow2      Deaths       66.64282   31.85692       15.893314
#>    overprediction coverage_deviation        bias   ae_median coverage_50
#> 1:   14096.100883        -0.11211957  0.09796875 38473.60156   0.3281250
#> 2:      65.899117         0.11614130  0.33906250   233.25781   0.6640625
#> 3:   10043.121943        -0.09785326 -0.05640625 24101.07031   0.3906250
#> 4:       7.138247         0.19528533  0.07265625    53.13281   0.8750000
#> 5:       8.978601        -0.02312500 -0.02234375    78.47656   0.4609375
#> 6:   11906.823030        -0.06660326 -0.07890625 27923.81250   0.4687500
#> 7:      18.892583        -0.04287176 -0.00512605   104.74790   0.4201681
#>    coverage_90
#> 1:   0.8203125
#> 2:   1.0000000
#> 3:   0.8046875
#> 4:   1.0000000
#> 5:   0.8750000
#> 6:   0.7890625
#> 7:   0.9075630

# forecast formats with different metrics
score(example_binary)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `true_value` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#> 2.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>      location location_name target_end_date target_type forecast_date
#>   1:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   2:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   3:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   4:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>   5:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>  ---                                                                 
#> 883:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 884:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 885:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 886:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 887:       IT         Italy      2021-07-24      Deaths    2021-07-12
#>                      model horizon prediction true_value brier_score log_score
#>   1: EuroCOVIDhub-ensemble       1      0.375          0    0.140625 0.4700036
#>   2: EuroCOVIDhub-baseline       1      0.475          0    0.225625 0.6443570
#>   3:  epiforecasts-EpiNow2       1      0.425          0    0.180625 0.5533852
#>   4: EuroCOVIDhub-ensemble       1      0.425          0    0.180625 0.5533852
#>   5: EuroCOVIDhub-baseline       1      0.500          1    0.250000 0.6931472
#>  ---                                                                          
#> 883: EuroCOVIDhub-baseline       2      0.250          0    0.062500 0.2876821
#> 884:       UMass-MechBayes       3      0.475          0    0.225625 0.6443570
#> 885:       UMass-MechBayes       2      0.450          0    0.202500 0.5978370
#> 886:  epiforecasts-EpiNow2       3      0.375          0    0.140625 0.4700036
#> 887:  epiforecasts-EpiNow2       2      0.300          0    0.090000 0.3566749
score(example_quantile)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>        location target_end_date target_type location_name forecast_date
#>     1:       DE      2021-05-08       Cases       Germany    2021-05-03
#>     2:       DE      2021-05-08       Cases       Germany    2021-05-03
#>     3:       DE      2021-05-08       Cases       Germany    2021-05-03
#>     4:       DE      2021-05-08       Cases       Germany    2021-05-03
#>     5:       DE      2021-05-08       Cases       Germany    2021-05-03
#>    ---                                                                 
#> 20397:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 20398:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 20399:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 20400:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 20401:       IT      2021-07-24      Deaths         Italy    2021-07-12
#>                        model horizon range interval_score dispersion
#>     1: EuroCOVIDhub-baseline       1     0       25620.00       0.00
#>     2: EuroCOVIDhub-baseline       1    10       25599.50     184.50
#>     3: EuroCOVIDhub-baseline       1    10       25599.50     184.50
#>     4: EuroCOVIDhub-baseline       1    20       25481.00     556.00
#>     5: EuroCOVIDhub-baseline       1    20       25481.00     556.00
#>    ---                                                              
#> 20397:  epiforecasts-EpiNow2       2    90          21.50      21.50
#> 20398:  epiforecasts-EpiNow2       2    95          13.90      13.90
#> 20399:  epiforecasts-EpiNow2       2    95          13.90      13.90
#> 20400:  epiforecasts-EpiNow2       2    98           6.71       6.71
#> 20401:  epiforecasts-EpiNow2       2    98           6.71       6.71
#>        underprediction overprediction coverage coverage_deviation bias quantile
#>     1:               0          25620        0               0.00 0.95    0.500
#>     2:               0          25415        0              -0.10 0.95    0.450
#>     3:               0          25415        0              -0.10 0.95    0.550
#>     4:               0          24925        0              -0.20 0.95    0.400
#>     5:               0          24925        0              -0.20 0.95    0.600
#>    ---                                                                         
#> 20397:               0              0        1               0.10 0.90    0.950
#> 20398:               0              0        1               0.05 0.90    0.025
#> 20399:               0              0        1               0.05 0.90    0.975
#> 20400:               0              0        1               0.02 0.90    0.010
#> 20401:               0              0        1               0.02 0.90    0.990
#>        ae_median quantile_coverage
#>     1:     25620              TRUE
#>     2:     25620              TRUE
#>     3:     25620              TRUE
#>     4:     25620              TRUE
#>     5:     25620              TRUE
#>    ---                            
#> 20397:       108              TRUE
#> 20398:       108             FALSE
#> 20399:       108              TRUE
#> 20400:       108             FALSE
#> 20401:       108              TRUE
score(example_integer)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>      location location_name target_end_date target_type forecast_date
#>   1:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   2:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   3:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   4:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>   5:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>  ---                                                                 
#> 883:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 884:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 885:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 886:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 887:       IT         Italy      2021-07-24      Deaths    2021-07-12
#>                      model horizon        mad  bias       dss        crps
#>   1: EuroCOVIDhub-ensemble       1 17640.7161  0.55 20.601389  7482.96187
#>   2: EuroCOVIDhub-baseline       1 19341.9996  0.95 22.443381 20371.25500
#>   3:  epiforecasts-EpiNow2       1 32348.8494  0.80 22.310571 24810.45938
#>   4: EuroCOVIDhub-ensemble       1   266.8680  0.20 11.112022    67.53063
#>   5: EuroCOVIDhub-baseline       1   397.3368 -0.05 11.541083    86.47313
#>  ---                                                                     
#> 883: EuroCOVIDhub-baseline       2   168.2751  0.45 12.913523    66.54000
#> 884:       UMass-MechBayes       3    25.2042  0.10  6.357686     6.64750
#> 885:       UMass-MechBayes       2    40.0302  0.80  8.769176    29.52313
#> 886:  epiforecasts-EpiNow2       3   108.2298  0.10 10.283440    28.84250
#> 887:  epiforecasts-EpiNow2       2    93.4038  0.65 10.563036    49.95750
#>      ae_median      se_mean
#>   1:    9414.0 2.303967e+08
#>   2:   29379.0 9.128526e+08
#>   3:   36513.0 1.603612e+09
#>   4:      77.0 8.588656e+03
#>   5:      27.5 2.118301e+03
#>  ---                       
#> 883:      58.5 6.063906e+04
#> 884:       5.0 6.084000e+01
#> 885:      42.5 2.159926e+03
#> 886:       8.0 3.214890e+03
#> 887:      73.5 1.262252e+04
score(example_continuous)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>      location location_name target_end_date target_type forecast_date
#>   1:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   2:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   3:       DE       Germany      2021-05-08       Cases    2021-05-03
#>   4:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>   5:       DE       Germany      2021-05-08      Deaths    2021-05-03
#>  ---                                                                 
#> 883:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 884:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 885:       IT         Italy      2021-07-24      Deaths    2021-07-12
#> 886:       IT         Italy      2021-07-24      Deaths    2021-07-05
#> 887:       IT         Italy      2021-07-24      Deaths    2021-07-12
#>                      model horizon         mad  bias       dss         crps
#>   1: EuroCOVIDhub-ensemble       1 17641.24334  0.55 20.601386  7482.975177
#>   2: EuroCOVIDhub-baseline       1 19341.68942  0.95 22.443370 20371.250988
#>   3:  epiforecasts-EpiNow2       1 32348.79978  0.80 22.310568 24810.424753
#>   4: EuroCOVIDhub-ensemble       1   267.13585  0.20 11.112388    67.510511
#>   5: EuroCOVIDhub-baseline       1   397.09371 -0.05 11.541040    86.462930
#>  ---                                                                       
#> 883: EuroCOVIDhub-baseline       2   168.03093  0.45 12.913561    66.515150
#> 884:       UMass-MechBayes       3    24.76457  0.10  6.353679     6.616381
#> 885:       UMass-MechBayes       2    39.96542  0.80  8.765716    29.446723
#> 886:  epiforecasts-EpiNow2       3   107.93293  0.10 10.283832    28.863742
#> 887:  epiforecasts-EpiNow2       2    93.54245  0.65 10.563729    49.998906
#>      log_score    ae_median      se_mean
#>   1: 10.989568  9413.839675 2.303980e+08
#>   2: 11.941135 29378.807513 9.128538e+08
#>   3: 12.007007 36512.810659 1.603609e+09
#>   4:  6.516954    77.077318 8.593954e+03
#>   5:  6.913879    27.429626 2.123067e+03
#>  ---                                    
#> 883:  6.067282    58.606302 6.065218e+04
#> 884:  4.242600     5.071742 6.043161e+01
#> 885:  5.321013    42.271396 2.153122e+03
#> 886:  5.709216     7.760566 3.212150e+03
#> 887:  5.736238    73.670518 1.263027e+04

# score point forecasts (marked by 'NA' in the quantile column)
score(example_point) %>%
  summarise_scores(by = "model", na.rm = TRUE)
#> The following messages were produced when checking inputs:
#> 1.  Some values for `prediction` are NA in the data provided and the corresponding rows were removed. This may indicate a problem if unexpected.
#>                    model interval_score dispersion underprediction
#> 1: EuroCOVIDhub-baseline     14092.7647 2192.26967      4996.92898
#> 2: EuroCOVIDhub-ensemble      8852.4196 1930.80064      2005.56357
#> 3:  epiforecasts-EpiNow2     10659.5125 3084.85850      1604.96135
#> 4:       UMass-MechBayes        51.4781   28.09387        15.32315
#>    overprediction coverage_deviation        bias    ae_point     se_point
#> 1:     6903.56605        0.002102273  0.23905983 19353.42969 2.883446e+09
#> 2:     4916.05540        0.050752841  0.01027027 12077.10156 1.945118e+09
#> 3:     5969.69268       -0.057861612 -0.04915929 14521.10526 2.680928e+09
#> 4:        8.06108       -0.024886364 -0.02847458    78.47656 1.170976e+04
#>    ae_median
#> 1:       NaN
#> 2:       NaN
#> 3:       NaN
#> 4:       NaN