Metrics for Binary Outcomes

Brier score

The Brier Score is the mean squared error between the probabilistic prediction and the observed outcome. The Brier score is a proper scoring rule. Small values are better (best is 0, the worst is 1).

$$ \textrm{Brier\_Score} = \frac{1}{N} \sum_{t = 1}^{n} (\textrm{prediction}_t - \textrm{outcome}_t)^2, $$ where \(\textrm{outcome}_t \in \{0, 1\}\), and \(\textrm{prediction}_t \in [0, 1]\) represents the probability that the outcome is equal to 1.

Log score for binary outcomes

The Log Score is the negative logarithm of the probability assigned to the observed value. It is a proper scoring rule. Small values are better (best is zero, worst is infinity).

Usage

brier_score(observed, predicted)

logs_binary(observed, predicted)

Arguments

observed

A factor of length n with exactly two levels, holding the observed values. The highest factor level is assumed to be the reference level. This means that predicted represents the probability that the observed value is equal to the highest factor level.

predicted

A numeric vector of length n, holding probabilities. Values represent the probability that the corresponding outcome is equal to the highest level of the factor observed.

Value

A numeric vector of size n with the Brier scores

A numeric vector of size n with log scores

Details

The functions require users to provide observed values as a factor in order to distinguish its input from the input format required for scoring point forecasts. Internally, however, factors will be converted to numeric values. A factor observed = factor(c(0, 1, 1, 0, 1) with two levels (0 and 1) would internally be coerced to a numeric vector (in this case this would result in the numeric vector c(1, 2, 2, 1, 1)). After subtracting 1, the resulting vector (c(0, 1, 1, 0) in this case) is used for internal calculations. All predictions are assumed represent the probability that the outcome is equal of the highest factor level (in this case that the outcome is equal to 1).

You could alternatively also provide a vector like observed = factor(c("a", "b", "b", "a")) (with two levels, a and b), which would result in exactly the same internal representation. Probabilities then represent the probability that the outcome is equal to "b". If you want your predictions to be probabilities that the outcome is "a", then you could of course make observed a factor with levels swapped, i.e. observed = factor(c("a", "b", "b", "a"), levels = c("b", "a"))

Examples

observed <- factor(sample(c(0, 1), size = 30, replace = TRUE))
predicted <- runif(n = 30, min = 0, max = 1)

brier_score(observed, predicted)
#>  [1] 1.995138e-01 2.393670e-01 5.623770e-02 3.804747e-01 4.981616e-01
#>  [6] 8.186614e-03 3.296334e-01 7.966937e-01 5.209740e-02 1.361491e-02
#> [11] 2.904885e-01 6.302358e-05 1.409077e-01 6.754322e-02 3.039031e-02
#> [16] 4.988815e-01 1.593357e-01 6.310935e-01 2.655075e-01 7.386602e-01
#> [21] 9.292061e-01 5.318306e-02 1.401406e-01 5.151116e-01 2.922472e-01
#> [26] 4.371923e-01 3.034565e-01 5.255338e-01 8.854089e-01 1.621539e-01
logs_binary(observed, predicted)
#>  [1] 0.591800113 0.671877833 0.270687163 0.959267000 1.223514670 0.094838190
#>  [7] 0.853637705 2.230981038 0.259092740 0.124071062 0.774291786 0.007970419
#> [13] 0.470606595 0.300957515 0.191557893 1.225248904 0.509441135 1.581890632
#> [19] 0.724171916 1.962216439 3.322942476 0.262163167 0.468969935 1.264830233
#> [25] 0.777831581 1.082360522 0.800440408 1.290755563 2.829569073 0.515308025