scoringutils

This package is designed to help with assessing the quality of predictions. It provides a collection of proper scoring rules and metrics as well that can be accessed independently or collectively through a higher-level wrapper function.

Predictions can be either probabilistic forecasts (generally predictive samples generated by Markov Chain Monte Carlo procedures), quantile forecasts or point forecasts. The true values can be either continuous, integer, or binary.

A collection of different metrics and scoring rules can be accessed through the function score(). Given a data.frame of the correct form the function will automatically figure out the type of prediction and true values and return appropriate scoring metrics.

The package also has a lot of default visualisation based on the output created by score().

plot_score_table()

plot_correlation()

plot_wis()

plot_ranges()

plot_heatmap()

plot_predictions()

plot_interval_coverage()

plot_quantile_coverage()

Alternatively, the following functions can be accessed directly:

brier_score()

pit()

bias_sample()

bias_quantile()

bias_range()

mad_sample()

crps_sample()

logs_sample()

dss_sample()

ae_median_sample()

Predictions can be evaluated in a lot of different formats. If you want to convert from one format to the other, the following helper functions can do that for you:

sample_to_quantile()