Squared Error of the Mean (Sample-based Version)

Squared error of the mean calculated as

$$ \text{mean}(\text{true_value} - \text{prediction})^2 $$

Usage

se_mean_sample(true_values, predictions)

Arguments

true_values

A vector with the true observed values of size n

predictions

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predictions can just be a vector of size n.

Value

vector with the scoring values

See also

Examples

true_values <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
se_mean_sample(true_values, predicted_values)
#>  [1]  2.463069049  0.186844064  0.537185991  0.004972353  0.513130043
#>  [6]  0.073721780  0.034274834  0.853252083  0.281044338  0.819898210
#> [11]  0.428003342 16.717567227  1.066823797  1.669540237  0.872621565
#> [16] 16.141430772  5.724808573  0.091218326  1.145746520  0.084491541
#> [21]  0.949368263  0.251567284  0.879409524  0.620822262  6.058534241
#> [26]  1.848901636  0.147761052  0.215228560  2.729704316  6.619693057