Squared error of the mean (sample-based version)

Squared error of the mean calculated as

$$ \textrm{mean}(\textrm{observed} - \textrm{mean prediction})^2 $$ The mean prediction is calculated as the mean of the predictive samples.

Usage

se_mean_sample(observed, predicted)

Arguments

observed

A vector with observed values of size n

predicted

nxN matrix of predictive samples, n (number of rows) being the number of data points and N (number of columns) the number of Monte Carlo samples. Alternatively, predicted can just be a vector of size n.

Examples

observed <- rnorm(30, mean = 1:30)
predicted_values <- matrix(rnorm(30, mean = 1:30))
se_mean_sample(observed, predicted_values)
#>  [1] 0.173184776 0.023435451 1.610298449 3.571463532 0.287637723 3.456505359
#>  [7] 3.104445200 6.416765323 0.014822594 0.262562121 0.002945155 2.290331532
#> [13] 0.062723285 0.492648216 0.047323098 1.489234187 1.959968436 2.103650766
#> [19] 0.086133601 0.894739782 5.993444771 2.978954352 6.778862480 1.779654975
#> [25] 1.712346282 0.843409936 3.193731460 0.514182655 0.134914452 0.485937390