Summarise scores as produced by score()

Summarise scores as produced by score().

summarise_scores relies on a way to identify the names of the scores and distinguish them from columns that denote the unit of a single forecast. Internally, this is done via a stored attribute, score_names that stores the names of the scores. This means, however, that you need to be careful with renaming scores after they have been produced by score(). If you do, you also have to manually update the attribute by calling attr(scores, "score_names") <- new_names.

Usage

summarise_scores(scores, by = "model", across = NULL, fun = mean, ...)

summarize_scores(scores, by = "model", across = NULL, fun = mean, ...)

Arguments

scores

A data.table of scores as produced by score().

by

character vector with column names to summarise scores by. Default is model, meaning that there will be one score per model in the output. The unit of a single forecast is determined by the columns present in the input data that do not correspond to a metric produced by score(), which indicate indicate a grouping of forecasts (for example there may be one forecast per day, location and model). Adding additional, unrelated, columns may alter results in an unpredictable way.

across

character vector with column names from the vector of variables that define the unit of a single forecast (see above) to summarise scores across (meaning that the specified columns will be dropped). This is an alternative to specifying by directly. If across is set, by will be ignored. If across is NULL (default), then by will be used.

fun

a function used for summarising scores. Default is mean.

...

additional parameters that can be passed to the summary function provided to fun. For more information see the documentation of the respective function.

Value

a data.table with summarised scores. Scores are summarised according to the names of the columns of the original data specified in by or across using the fun passed to summarise_scores().

Examples

# \dontshow{
  data.table::setDTthreads(2) # restricts number of cores used on CRAN
# }
library(magrittr) # pipe operator
scores <- score(example_continuous)
#> ℹ Some rows containing NA values may be removed. This is fine if not
#>   unexpected.

# get scores by model
summarise_scores(scores,by = "model")
#>                    model         bias      dss        crps log_score        mad
#>                   <char>        <num>    <num>       <num>     <num>      <num>
#> 1: EuroCOVIDhub-ensemble  0.009765625 16.40496  9876.95886 10.747811  8763.6176
#> 2: EuroCOVIDhub-baseline  0.177734375      NaN 15309.68627       Inf  9680.3792
#> 3:  epiforecasts-EpiNow2 -0.024898785 26.10137 11901.43354       Inf 12999.5404
#> 4:       UMass-MechBayes -0.026953125 10.08582    60.19018  5.941622   123.6211
#>      ae_median      se_mean
#>          <num>        <num>
#> 1: 12406.03100 2.103026e+09
#> 2: 18932.50196 2.885063e+09
#> 3: 14680.12285 3.152268e+09
#> 4:    79.66001 1.371418e+04

# get scores by model and target type
summarise_scores(scores, by = c("model", "target_type"))
#>                    model target_type        bias      dss        crps log_score
#>                   <char>      <char>       <num>    <num>       <num>     <num>
#> 1: EuroCOVIDhub-ensemble       Cases -0.04648437 22.89997 19703.05522 15.633420
#> 2: EuroCOVIDhub-baseline       Cases  0.03671875      NaN 30453.58346       Inf
#> 3:  epiforecasts-EpiNow2       Cases -0.03867188 40.87716 22896.51608       Inf
#> 4: EuroCOVIDhub-ensemble      Deaths  0.06601562  9.90995    50.86249  5.862203
#> 5: EuroCOVIDhub-baseline      Deaths  0.31875000 12.99360   165.78907  6.977391
#> 6:       UMass-MechBayes      Deaths -0.02695313 10.08582    60.19018  5.941622
#> 7:  epiforecasts-EpiNow2      Deaths -0.01008403 10.20807    74.79013  6.024092
#>           mad   ae_median      se_mean
#>         <num>       <num>        <num>
#> 1: 17385.2629 24749.39707 4.206042e+09
#> 2: 18982.2128 37648.01693 5.769964e+09
#> 3: 24929.3438 28233.04536 6.082863e+09
#> 4:   141.9723    62.66492 1.080233e+04
#> 5:   378.5457   216.98699 1.622417e+05
#> 6:   123.6211    79.66001 1.371418e+04
#> 7:   167.4829   102.18939 3.243111e+04

# Get scores summarised across horizon, forecast date, and target end date
summarise_scores(
 scores, across = c("horizon", "forecast_date", "target_end_date")
)
#>     location  location_name target_type                 model        bias
#>       <char>         <char>      <char>                <char>       <num>
#>  1:       DE        Germany       Cases EuroCOVIDhub-ensemble  0.17812500
#>  2:       DE        Germany       Cases EuroCOVIDhub-baseline  0.51093750
#>  3:       DE        Germany       Cases  epiforecasts-EpiNow2  0.15937500
#>  4:       DE        Germany      Deaths EuroCOVIDhub-ensemble -0.04531250
#>  5:       DE        Germany      Deaths EuroCOVIDhub-baseline  0.40312500
#>  6:       DE        Germany      Deaths       UMass-MechBayes -0.58125000
#>  7:       DE        Germany      Deaths  epiforecasts-EpiNow2 -0.34531250
#>  8:       FR         France       Cases EuroCOVIDhub-ensemble  0.01250000
#>  9:       FR         France       Cases EuroCOVIDhub-baseline -0.02968750
#> 10:       FR         France       Cases  epiforecasts-EpiNow2 -0.02656250
#> 11:       FR         France      Deaths EuroCOVIDhub-ensemble  0.04687500
#> 12:       FR         France      Deaths EuroCOVIDhub-baseline  0.34218750
#> 13:       FR         France      Deaths       UMass-MechBayes  0.05156250
#> 14:       FR         France      Deaths  epiforecasts-EpiNow2 -0.05652174
#> 15:       GB United Kingdom       Cases EuroCOVIDhub-ensemble -0.47656250
#> 16:       GB United Kingdom       Cases EuroCOVIDhub-baseline -0.61875000
#> 17:       GB United Kingdom       Cases  epiforecasts-EpiNow2 -0.18750000
#> 18:       GB United Kingdom      Deaths EuroCOVIDhub-ensemble  0.10937500
#> 19:       GB United Kingdom      Deaths EuroCOVIDhub-baseline  0.08281250
#> 20:       GB United Kingdom      Deaths       UMass-MechBayes  0.62812500
#> 21:       GB United Kingdom      Deaths  epiforecasts-EpiNow2  0.43125000
#> 22:       IT          Italy       Cases EuroCOVIDhub-ensemble  0.10000000
#> 23:       IT          Italy       Cases EuroCOVIDhub-baseline  0.28437500
#> 24:       IT          Italy       Cases  epiforecasts-EpiNow2 -0.10000000
#> 25:       IT          Italy      Deaths EuroCOVIDhub-ensemble  0.15312500
#> 26:       IT          Italy      Deaths EuroCOVIDhub-baseline  0.44687500
#> 27:       IT          Italy      Deaths       UMass-MechBayes -0.20625000
#> 28:       IT          Italy      Deaths  epiforecasts-EpiNow2 -0.08281250
#>     location  location_name target_type                 model        bias
#>           dss        crps log_score         mad   ae_median      se_mean
#>         <num>       <num>     <num>       <num>       <num>        <num>
#>  1: 18.913810  7936.62666 10.368232 11207.25397 11618.84123 4.494508e+08
#>  2: 21.878467 15588.64493 11.938333 13450.95537 20043.14918 9.339773e+08
#>  3: 29.470535 13453.07861 22.686464 14195.91295 18498.57965 1.286406e+09
#>  4: 10.257710    56.54768  6.104505   167.78914    67.75772 1.061735e+04
#>  5: 12.536674   155.04973  6.997312   391.89252   219.90763 1.141223e+05
#>  6: 10.620902    76.81642  6.345002   147.52889   113.14713 1.543732e+04
#>  7: 11.020430    98.55426  6.571691   222.78911   140.73457 3.401571e+04
#>  8: 31.176915 47132.62230 28.508977 20503.83242 54298.94987 1.427192e+10
#>  9:       NaN 56093.15772       Inf 25117.85238 65048.92899 1.653529e+10
#> 10: 34.493853 54117.03193 32.355909 33004.99213 63535.47737 1.981982e+10
#> 11: 10.729140    67.93297  6.236167   188.67948    85.67433 1.819764e+04
#> 12: 13.289991   202.77900  7.204643   501.10175   259.38228 2.316930e+05
#> 13: 11.322153    85.42422  6.518248   186.16678    98.33634 2.700122e+04
#> 14: 11.196840   119.95348  6.553617   331.22963   160.27973 8.603595e+04
#> 15: 21.106923 17824.87005 11.748949 30723.25562 24247.60751 1.940949e+09
#> 16: 24.877031 40848.86212 13.554332 22622.32113 51882.49530 5.228815e+09
#> 17: 21.224165 20508.11818 11.483762 47677.95848 25915.83258 3.160764e+09
#> 18:  8.839111    25.48464  5.251438    81.18266    26.29047 1.515431e+03
#> 19: 13.061990   109.66110  6.625700   212.20871   110.49682 1.048123e+05
#> 20:  9.040336    40.55896  5.310473    67.03450    62.02151 8.933439e+03
#> 21:  8.859818    29.31789  5.233896    73.15706    35.78943 4.412517e+03
#> 22: 20.402224  5918.10189 11.907522  7106.70962  8832.18969 1.618505e+08
#> 23: 20.338858  9283.66908 10.903144 14737.72225 13617.49426 3.817765e+08
#> 24: 78.320096  3507.83559       Inf  4838.51176  4982.29184 6.446300e+07
#> 25:  9.813839    53.48466  5.856701   130.23793    70.93715 1.287888e+04
#> 26: 13.085745   195.66648  7.081908   408.97976   278.16121 1.983392e+05
#> 27:  9.359876    37.96112  5.592767    93.75413    45.13505 3.484721e+03
#> 28: 10.033302    64.03710  5.886091    88.80946    88.29173 2.033662e+04
#>           dss        crps log_score         mad   ae_median      se_mean

# get standard deviation
summarise_scores(scores, by = "model", fun = sd)
#>                    model      bias        dss        crps log_score        mad
#>                   <char>     <num>      <num>       <num>     <num>      <num>
#> 1: EuroCOVIDhub-ensemble 0.5468290  14.869520 39368.24836 21.510119 19799.1620
#> 2: EuroCOVIDhub-baseline 0.5457971         NA 45020.82814       NaN 13610.4174
#> 3:  epiforecasts-EpiNow2 0.6083410 108.130107 44957.07746       NaN 29616.1714
#> 4:       UMass-MechBayes 0.6221914   2.248998    49.62465  1.126019   123.3465
#>      ae_median      se_mean
#>          <num>        <num>
#> 1: 42801.64123 1.564286e+10
#> 2: 49458.36446 1.760651e+10
#> 3: 51129.54601 2.209086e+10
#> 4:    76.09471 2.994664e+04

# round digits
summarise_scores(scores,by = "model") %>%
  summarise_scores(fun = signif, digits = 2)
#>                    model    bias   dss  crps log_score   mad ae_median se_mean
#>                   <char>   <num> <num> <num>     <num> <num>     <num>   <num>
#> 1: EuroCOVIDhub-ensemble  0.0098    16  9900      11.0  8800     12000 2.1e+09
#> 2: EuroCOVIDhub-baseline  0.1800   NaN 15000       Inf  9700     19000 2.9e+09
#> 3:  epiforecasts-EpiNow2 -0.0250    26 12000       Inf 13000     15000 3.2e+09
#> 4:       UMass-MechBayes -0.0270    10    60       5.9   120        80 1.4e+04