Summarise scores as produced by score()

Summarise scores as produced by score()

Usage

summarise_scores(scores, by = NULL, across = NULL, fun = mean, ...)

summarize_scores(scores, by = NULL, across = NULL, fun = mean, ...)

Arguments

scores

A data.table of scores as produced by score().

by

character vector with column names to summarise scores by. Default is NULL, meaning that the only summary that takes is place is summarising over samples or quantiles (in case of quantile-based forecasts), such that there is one score per forecast as defined by the unit of a single forecast (rather than one score for every sample or quantile). The unit of a single forecast is determined by the columns present in the input data that do not correspond to a metric produced by score(), which indicate indicate a grouping of forecasts (for example there may be one forecast per day, location and model). Adding additional, unrelated, columns may alter results in an unpredictable way.

across

character vector with column names from the vector of variables that define the unit of a single forecast (see above) to summarise scores across (meaning that the specified columns will be dropped). This is an alternative to specifying by directly. If NULL (default), then by will be used or inferred internally if also not specified. Only one of across and by may be used at a time.

fun

a function used for summarising scores. Default is mean.

...

additional parameters that can be passed to the summary function provided to fun. For more information see the documentation of the respective function.

Value

a data.table with summarised scores. Scores are summarised according to the names of the columns of the original data specified in by or across using the fun passed to summarise_scores().

Examples

# \dontshow{
  data.table::setDTthreads(2) # restricts number of cores used on CRAN
# }
library(magrittr) # pipe operator
if (FALSE) {
scores <- score(example_continuous)
summarise_scores(scores)
}

# summarise over samples or quantiles to get one score per forecast
scores <- score(example_quantile)
#> Some rows containing NA values may be removed. This is fine if not unexpected.
summarise_scores(scores)
#>      location target_end_date target_type location_name forecast_date
#>        <char>          <Date>      <char>        <char>        <Date>
#>   1:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   2:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   3:       DE      2021-05-08       Cases       Germany    2021-05-03
#>   4:       DE      2021-05-08      Deaths       Germany    2021-05-03
#>   5:       DE      2021-05-08      Deaths       Germany    2021-05-03
#>  ---                                                                 
#> 883:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 884:       IT      2021-07-24      Deaths         Italy    2021-07-05
#> 885:       IT      2021-07-24      Deaths         Italy    2021-07-12
#> 886:       IT      2021-07-24      Deaths         Italy    2021-07-05
#> 887:       IT      2021-07-24      Deaths         Italy    2021-07-12
#>                      model horizon          wis overprediction underprediction
#>                     <char>   <num>        <num>          <num>           <num>
#>   1: EuroCOVIDhub-ensemble       1  7990.854783   2.549870e+03       0.0000000
#>   2: EuroCOVIDhub-baseline       1 16925.046957   1.527583e+04       0.0000000
#>   3:  epiforecasts-EpiNow2       1 25395.960870   1.722226e+04       0.0000000
#>   4: EuroCOVIDhub-ensemble       1    53.880000   0.000000e+00       0.6086957
#>   5: EuroCOVIDhub-baseline       1    46.793043   2.130435e+00       0.0000000
#>  ---                                                                          
#> 883: EuroCOVIDhub-baseline       2    80.336957   3.608696e+00       0.0000000
#> 884:       UMass-MechBayes       3     4.881739   4.347826e-02       0.0000000
#> 885:       UMass-MechBayes       2    25.581739   1.782609e+01       0.0000000
#> 886:  epiforecasts-EpiNow2       3    19.762609   5.478261e+00       0.0000000
#> 887:  epiforecasts-EpiNow2       2    66.161739   4.060870e+01       0.0000000
#>       dispersion  bias interval_coverage_50 interval_coverage_90
#>            <num> <num>                <num>                <num>
#>   1: 5440.985217  0.50                    1                    1
#>   2: 1649.220870  0.95                    0                    0
#>   3: 8173.700000  0.90                    0                    1
#>   4:   53.271304 -0.10                    1                    1
#>   5:   44.662609  0.30                    1                    1
#>  ---                                                            
#> 883:   76.728261  0.20                    1                    1
#> 884:    4.838261  0.10                    1                    1
#> 885:    7.755652  0.80                    0                    1
#> 886:   14.284348  0.50                    1                    1
#> 887:   25.553043  0.90                    0                    1
#>      interval_coverage_deviation ae_median
#>                            <num>     <num>
#>   1:                  0.05181818     12271
#>   2:                 -0.40272727     25620
#>   3:                 -0.31181818     44192
#>   4:                  0.41545455        14
#>   5:                  0.23363636        15
#>  ---                                      
#> 883:                  0.32454545        53
#> 884:                  0.41545455         1
#> 885:                 -0.22090909        46
#> 886:                  0.05181818        26
#> 887:                 -0.31181818       108

# get scores by model
summarise_scores(scores,by = "model")
#>                    model         wis overprediction underprediction dispersion
#>                   <char>       <num>          <num>           <num>      <num>
#> 1: EuroCOVIDhub-ensemble  8992.62316    5025.130095      2120.64029 1846.85278
#> 2: EuroCOVIDhub-baseline 14321.48926    7081.000000      5143.53567 2096.95360
#> 3:  epiforecasts-EpiNow2 10827.40786    6179.439535      1697.23411 2950.73422
#> 4:       UMass-MechBayes    52.65195       8.978601        16.80095   26.87239
#>           bias interval_coverage_50 interval_coverage_90
#>          <num>                <num>                <num>
#> 1:  0.00812500            0.6328125            0.9023438
#> 2:  0.21851562            0.4960938            0.9101562
#> 3: -0.04336032            0.4453441            0.8461538
#> 4: -0.02234375            0.4609375            0.8750000
#>    interval_coverage_deviation   ae_median
#>                          <num>       <num>
#> 1:                 0.050752841 12077.10156
#> 2:                 0.002102273 19353.42969
#> 3:                -0.057861612 14521.10526
#> 4:                -0.024886364    78.47656

# get scores by model and target type
summarise_scores(scores, by = c("model", "target_type"))
#>                    model target_type         wis overprediction underprediction
#>                   <char>      <char>       <num>          <num>           <num>
#> 1: EuroCOVIDhub-ensemble       Cases 17943.82383   10043.121943     4237.177310
#> 2: EuroCOVIDhub-baseline       Cases 28483.57465   14096.100883    10284.972826
#> 3:  epiforecasts-EpiNow2       Cases 20831.55662   11906.823030     3260.355639
#> 4: EuroCOVIDhub-ensemble      Deaths    41.42249       7.138247        4.103261
#> 5: EuroCOVIDhub-baseline      Deaths   159.40387      65.899117        2.098505
#> 6:       UMass-MechBayes      Deaths    52.65195       8.978601       16.800951
#> 7:  epiforecasts-EpiNow2      Deaths    66.64282      18.892583       15.893314
#>    dispersion        bias interval_coverage_50 interval_coverage_90
#>         <num>       <num>                <num>                <num>
#> 1: 3663.52458 -0.05640625            0.3906250            0.8046875
#> 2: 4102.50094  0.09796875            0.3281250            0.8203125
#> 3: 5664.37795 -0.07890625            0.4687500            0.7890625
#> 4:   30.18099  0.07265625            0.8750000            1.0000000
#> 5:   91.40625  0.33906250            0.6640625            1.0000000
#> 6:   26.87239 -0.02234375            0.4609375            0.8750000
#> 7:   31.85692 -0.00512605            0.4201681            0.9075630
#>    interval_coverage_deviation   ae_median
#>                          <num>       <num>
#> 1:                 -0.10230114 24101.07031
#> 2:                 -0.11721591 38473.60156
#> 3:                 -0.06963068 27923.81250
#> 4:                  0.20380682    53.13281
#> 5:                  0.12142045   233.25781
#> 6:                 -0.02488636    78.47656
#> 7:                 -0.04520244   104.74790

# Get scores summarised across horizon, forecast date, and target end date
summarise_scores(
 scores, across = c("horizon", "forecast_date", "target_end_date")
)
#>     location target_type  location_name                 model         wis
#>       <char>      <char>         <char>                <char>       <num>
#>  1:       DE       Cases        Germany EuroCOVIDhub-ensemble  6286.66495
#>  2:       DE       Cases        Germany EuroCOVIDhub-baseline 14506.65500
#>  3:       DE       Cases        Germany  epiforecasts-EpiNow2 11684.72865
#>  4:       DE      Deaths        Germany EuroCOVIDhub-ensemble    44.46077
#>  5:       DE      Deaths        Germany EuroCOVIDhub-baseline   155.91235
#>  6:       DE      Deaths        Germany       UMass-MechBayes    68.91583
#>  7:       DE      Deaths        Germany  epiforecasts-EpiNow2    93.33921
#>  8:       FR       Cases         France EuroCOVIDhub-ensemble 44537.04769
#>  9:       FR       Cases         France EuroCOVIDhub-baseline 54147.68308
#> 10:       FR       Cases         France  epiforecasts-EpiNow2 50141.70361
#> 11:       FR      Deaths         France EuroCOVIDhub-ensemble    56.70677
#> 12:       FR      Deaths         France EuroCOVIDhub-baseline   187.63842
#> 13:       FR      Deaths         France       UMass-MechBayes    74.17250
#> 14:       FR      Deaths         France  epiforecasts-EpiNow2    96.69760
#> 15:       GB       Cases United Kingdom EuroCOVIDhub-ensemble 16010.55516
#> 16:       GB       Cases United Kingdom EuroCOVIDhub-baseline 36032.52912
#> 17:       GB       Cases United Kingdom  epiforecasts-EpiNow2 18303.21825
#> 18:       GB      Deaths United Kingdom EuroCOVIDhub-ensemble    20.04649
#> 19:       GB      Deaths United Kingdom EuroCOVIDhub-baseline    95.17583
#> 20:       GB      Deaths United Kingdom       UMass-MechBayes    33.45330
#> 21:       GB      Deaths United Kingdom  epiforecasts-EpiNow2    25.77830
#> 22:       IT       Cases          Italy EuroCOVIDhub-ensemble  4941.02753
#> 23:       IT       Cases          Italy EuroCOVIDhub-baseline  9247.43141
#> 24:       IT       Cases          Italy  epiforecasts-EpiNow2  3196.57595
#> 25:       IT      Deaths          Italy EuroCOVIDhub-ensemble    44.47594
#> 26:       IT      Deaths          Italy EuroCOVIDhub-baseline   198.88887
#> 27:       IT      Deaths          Italy       UMass-MechBayes    34.06615
#> 28:       IT      Deaths          Italy  epiforecasts-EpiNow2    59.20908
#>     location target_type  location_name                 model         wis
#>     overprediction underprediction  dispersion        bias interval_coverage_50
#>              <num>           <num>       <num>       <num>                <num>
#>  1:   3.522946e+03    4.768886e+02  2286.83071  0.17375000            0.4062500
#>  2:   1.146823e+04    1.147921e+02  2923.63054  0.49062500            0.3750000
#>  3:   7.382238e+03    9.200027e+02  3382.48817  0.13375000            0.3750000
#>  4:   4.653533e+00    4.533967e+00    35.27327 -0.06875000            0.9062500
#>  5:   7.764538e+01    0.000000e+00    78.26697  0.52812500            0.6250000
#>  6:   0.000000e+00    3.715761e+01    31.75822 -0.63312500            0.3125000
#>  7:   1.092799e+01    3.803397e+01    44.37726 -0.45937500            0.3437500
#>  8:   3.364876e+04    6.054179e+03  4834.10747  0.02281250            0.3750000
#>  9:   3.927517e+04    8.578344e+03  6294.16678  0.16250000            0.2812500
#> 10:   3.500782e+04    7.281398e+03  7852.48758 -0.05718750            0.5312500
#> 11:   7.135870e+00    7.942935e+00    41.62796  0.06562500            0.9375000
#> 12:   8.644158e+01    0.000000e+00   101.19685  0.46562500            0.5312500
#> 13:   1.265761e+01    2.091984e+01    40.59505  0.08812500            0.5937500
#> 14:   2.576938e+01    1.061815e+01    60.31008 -0.07173913            0.4782609
#> 15:   1.102159e+03    8.891671e+03  6016.72500 -0.50937500            0.3437500
#> 16:   6.210326e+01    3.166997e+04  4300.45982 -0.63156250            0.2500000
#> 17:   4.294447e+03    3.617322e+03 10391.44923 -0.28125000            0.5625000
#> 18:   2.604620e+00    1.241848e+00    16.20003  0.15937500            0.8750000
#> 19:   5.203804e-01    8.266304e+00    86.38914 -0.10625000            0.9687500
#> 20:   1.905842e+01    4.891304e-02    14.34596  0.64468750            0.3125000
#> 21:   1.205163e+01    1.086957e-02    13.71580  0.54875000            0.4062500
#> 22:   1.898622e+03    1.525970e+03  1516.43514  0.08718750            0.4375000
#> 23:   5.578895e+03    7.767894e+02  2891.74663  0.37031250            0.4062500
#> 24:   9.427894e+02    1.222700e+03  1031.08682 -0.11093750            0.4062500
#> 25:   1.415897e+01    2.694293e+00    27.62268  0.13437500            0.7812500
#> 26:   9.898913e+01    1.277174e-01    99.77202  0.46875000            0.5312500
#> 27:   4.198370e+00    9.077446e+00    20.79034 -0.18906250            0.6250000
#> 28:   2.875543e+01    1.342663e+01    17.02701 -0.05687500            0.4687500
#>     overprediction underprediction  dispersion        bias interval_coverage_50
#>     interval_coverage_90 interval_coverage_deviation   ae_median
#>                    <num>                       <num>       <num>
#>  1:            0.8750000                 -0.06465909  9949.12500
#>  2:            0.7812500                 -0.09022727 21109.81250
#>  3:            0.6875000                 -0.15272727 18315.09375
#>  4:            1.0000000                  0.24784091    51.93750
#>  5:            1.0000000                  0.02625000   261.03125
#>  6:            0.9375000                 -0.07602273   124.50000
#>  7:            1.0000000                 -0.06465909   154.62500
#>  8:            0.7500000                 -0.14136364 52432.93750
#>  9:            0.8125000                 -0.14136364 67726.87500
#> 10:            0.7187500                 -0.06750000 60549.84375
#> 11:            1.0000000                  0.21375000    72.12500
#> 12:            1.0000000                  0.08306818   293.59375
#> 13:            0.8750000                  0.05750000    88.53125
#> 14:            0.9565217                  0.02415020   143.00000
#> 15:            0.8125000                 -0.09875000 26422.84375
#> 16:            0.7187500                 -0.18965909 49934.03125
#> 17:            0.9375000                  0.01772727 27901.06250
#> 18:            1.0000000                  0.20238636    22.96875
#> 19:            1.0000000                  0.31318182    59.03125
#> 20:            0.7812500                 -0.13568182    55.34375
#> 21:            0.9375000                 -0.01920455    40.62500
#> 22:            0.7812500                 -0.10443182  7599.37500
#> 23:            0.9687500                 -0.04761364 15123.68750
#> 24:            0.8125000                 -0.07602273  4929.25000
#> 25:            1.0000000                  0.15125000    65.50000
#> 26:            1.0000000                  0.06318182   319.37500
#> 27:            0.9062500                  0.05465909    45.53125
#> 28:            0.7500000                 -0.10159091    91.50000
#>     interval_coverage_90 interval_coverage_deviation   ae_median

# get standard deviation
summarise_scores(scores, by = "model", fun = sd)
#>                    model       wis overprediction underprediction dispersion
#>                   <char>     <num>          <num>           <num>      <num>
#> 1: EuroCOVIDhub-ensemble 38788.740     36978.7730       7179.5954  3863.3379
#> 2: EuroCOVIDhub-baseline 43157.488     38723.6131      18417.2012  2682.7722
#> 3:  epiforecasts-EpiNow2 42478.208     38832.5598       7629.9180  6698.7470
#> 4:       UMass-MechBayes    42.813        18.3498         30.8519    25.2045
#>         bias interval_coverage_50 interval_coverage_90
#>        <num>                <num>                <num>
#> 1: 0.5707267            0.4829825            0.2974308
#> 2: 0.5628605            0.5009641            0.2865179
#> 3: 0.6632842            0.4980129            0.3615338
#> 4: 0.6457779            0.5004304            0.3320184
#>    interval_coverage_deviation   ae_median
#>                          <num>       <num>
#> 1:                   0.2848161 42500.79359
#> 2:                   0.2707257 50186.94967
#> 3:                   0.2964815 49800.66917
#> 4:                   0.2816340    74.79904

# round digits
summarise_scores(scores,by = "model") %>%
  summarise_scores(fun = signif, digits = 2)
#>                    model   wis overprediction underprediction dispersion
#>                   <char> <num>          <num>           <num>      <num>
#> 1: EuroCOVIDhub-ensemble  9000           5000            2100       1800
#> 2: EuroCOVIDhub-baseline 14000           7100            5100       2100
#> 3:  epiforecasts-EpiNow2 11000           6200            1700       3000
#> 4:       UMass-MechBayes    53              9              17         27
#>       bias interval_coverage_50 interval_coverage_90
#>      <num>                <num>                <num>
#> 1:  0.0081                 0.63                 0.90
#> 2:  0.2200                 0.50                 0.91
#> 3: -0.0430                 0.45                 0.85
#> 4: -0.0220                 0.46                 0.88
#>    interval_coverage_deviation ae_median
#>                          <num>     <num>
#> 1:                      0.0510     12000
#> 2:                      0.0021     19000
#> 3:                     -0.0580     15000
#> 4:                     -0.0250        78

# get quantiles of scores
# make sure to aggregate over ranges first
summarise_scores(scores,
  by = "model", fun = quantile,
  probs = c(0.25, 0.5, 0.75)
)
#>                     model         wis overprediction underprediction dispersion
#>                    <char>       <num>          <num>           <num>      <num>
#>  1: EuroCOVIDhub-ensemble    28.33707      0.0000000      0.00000000   23.67087
#>  2: EuroCOVIDhub-ensemble   300.30739      0.2826087      0.00000000  168.95543
#>  3: EuroCOVIDhub-ensemble  5532.28565     43.5217391     44.10869565 2030.41870
#>  4: EuroCOVIDhub-baseline   130.23793      0.0000000      0.00000000   87.81804
#>  5: EuroCOVIDhub-baseline   934.41870     54.1304348      0.00000000  163.28739
#>  6: EuroCOVIDhub-baseline 10795.61967    496.7065217      4.42391304 3476.52750
#>  7:  epiforecasts-EpiNow2    50.94935      0.0000000      0.00000000   16.90326
#>  8:  epiforecasts-EpiNow2   442.81696      0.0000000      0.08695652  152.12565
#>  9:  epiforecasts-EpiNow2  5902.85957     45.6956522    155.76086957 2033.24370
#> 10:       UMass-MechBayes    18.49435      0.0000000      0.00000000    7.73837
#> 11:       UMass-MechBayes    46.05413      0.0000000      0.02173913   17.15413
#> 12:       UMass-MechBayes    73.27087     10.2500000     18.41304348   38.04620
#>      bias interval_coverage_50 interval_coverage_90 interval_coverage_deviation
#>     <num>                <num>                <num>                       <num>
#>  1: -0.40                    0                    1                 -0.13000000
#>  2:  0.10                    1                    1                  0.14272727
#>  3:  0.40                    1                    1                  0.25636364
#>  4: -0.20                    0                    1                 -0.13000000
#>  5:  0.40                    0                    1                 -0.03909091
#>  6:  0.60                    1                    1                  0.23363636
#>  7: -0.70                    0                    1                 -0.31181818
#>  8: -0.10                    0                    1                 -0.03909091
#>  9:  0.60                    1                    1                  0.14272727
#> 10: -0.60                    0                    1                 -0.22090909
#> 11: -0.05                    0                    1                 -0.03909091
#> 12:  0.60                    1                    1                  0.23363636
#>     ae_median
#>         <num>
#>  1:     27.00
#>  2:    277.00
#>  3:   8100.00
#>  4:    166.00
#>  5:    763.50
#>  6:  17709.00
#>  7:     68.00
#>  8:    404.00
#>  9:   7177.00
#> 10:     24.75
#> 11:     52.00
#> 12:    106.00

# get ranges
# summarise_scores(scores, by = "range")